+++ src/core/dev/dpcp/hw_queue_tx_dpcp.cpp
@@ -31,21 +31,22 @@
  * SOFTWARE.
  */
 
+#include "config.h"
+#ifdef DEFINED_DPCP_PATH_TX
 #include <algorithm>
 #include <sys/mman.h>
 #include <netinet/ip.h>
-#include "dev/hw_queue_tx.h"
+#include "dev/dpcp/hw_queue_tx_dpcp.h"
 #include "dev/ring_simple.h"
 #include "proto/tls.h"
 #include "util/valgrind.h"
-#include <doca_buf.h>
-#include <doca_log.h>
 #include <thread>
 #include <cinttypes>
 #include <sock/sock-app.h>
+#include "dev/dpcp/cq_mgr_rx_regrq.h"
 
 #undef MODULE_NAME
-#define MODULE_NAME "hw_queue_tx"
+#define MODULE_NAME "hw_queue_tx_dpcp"
 DOCA_LOG_REGISTER(hw_queue_tx);
 
 #define hwqtx_logpanic   __log_info_panic
@@ -55,20 +56,6 @@ DOCA_LOG_REGISTER(hw_queue_tx);
 #define hwqtx_logdbg     __log_info_dbg
 #define hwqtx_logfunc    __log_info_func
 #define hwqtx_logfuncall __log_info_funcall
-
-#if !defined(MLX5_ETH_INLINE_HEADER_SIZE)
-#define MLX5_ETH_INLINE_HEADER_SIZE 18
-#endif
-
-#define OCTOWORD                   16
-#define WQEBB                      64
-#define DOCA_EXPAND_BATCH_SIZE     (1024)
-#define DOCA_MAX_LSO_HEADER        (64)
-#define DOCA_CHECKSUM_HW_L3_ENABLE (1)
-#define DOCA_CHECKSUM_HW_L4_ENABLE (1)
-#define DOCA_MAX_SGE_WITHOUT_TSO   (1)
-
-//#define DBG_DUMP_WQE 1
 
 #ifdef DBG_DUMP_WQE
 #define dbg_dump_wqe(_addr, _size)                                                                 \
@@ -84,7 +71,11 @@ DOCA_LOG_REGISTER(hw_queue_tx);
 #define dbg_dump_wqe(_addr, _size)
 #endif
 
-lso_metadata_pool *g_lso_metadata_pool = nullptr;
+#if !defined(MLX5_ETH_INLINE_HEADER_SIZE)
+#define MLX5_ETH_INLINE_HEADER_SIZE 18
+#endif
+#define OCTOWORD 16
+#define WQEBB    64
 
 static inline uint64_t align_to_octoword_up(uint64_t val)
 {
@@ -134,20 +125,15 @@ static inline uint32_t get_mlx5_opcode(xlio_ibv_wr_opcode verbs_opcode)
 
 hw_queue_tx::hw_queue_tx(ring_simple *ring, const slave_data_t *slave,
                          struct ibv_comp_channel *p_tx_comp_event_channel, const uint32_t tx_num_wr)
-    : m_p_ring(ring)
-    , m_p_ib_ctx_handler(slave->p_ib_ctx)
-    , m_n_sysvar_tx_num_wr_to_signal(safe_mce_sys().tx_num_wr_to_signal)
+    : m_n_sysvar_tx_num_wr_to_signal(safe_mce_sys().tx_num_wr_to_signal)
     , m_tx_num_wr(tx_num_wr)
     , m_port_num(slave->port_num)
-    , m_doca_mmap(g_buffer_pool_tx->get_doca_mmap())
+    , m_p_ring(ring)
+    , m_p_ib_ctx_handler(slave->p_ib_ctx)
 {
     hwqtx_logfunc(LOG_FUNCTION_CALL);
 
     memset(&m_hwq_tx_stats, 0, sizeof(m_hwq_tx_stats));
-
-    if (!prepare_doca_txq()) {
-        throw_xlio_exception("Failed to create DOCA TXQ");
-    }
 
     memset(&m_mlx5_qp, 0, sizeof(m_mlx5_qp));
 
@@ -169,8 +155,6 @@ hw_queue_tx::hw_queue_tx(ring_simple *ring, const slave_data_t *slave,
 hw_queue_tx::~hw_queue_tx()
 {
     hwqtx_logfunc(LOG_FUNCTION_CALL);
-
-    m_doca_txq.reset(nullptr); // Must be destroyed before TX PE.
 
     if (m_sq_wqe_idx_to_prop) {
         if (0 != munmap(m_sq_wqe_idx_to_prop, m_tx_num_wr * sizeof(*m_sq_wqe_idx_to_prop))) {
@@ -198,257 +182,7 @@ hw_queue_tx::~hw_queue_tx()
         m_p_cq_mgr_tx = nullptr;
     }
 
-    if (m_p_doca_lso_metadata_list) {
-        g_lso_metadata_pool->put_objs(m_p_doca_lso_metadata_list);
-    }
-
     hwqtx_logdbg("Destructor hw_queue_tx end");
-}
-
-bool hw_queue_tx::check_doca_caps(doca_devinfo *devinfo, uint32_t &max_burst_size,
-                                  uint32_t &max_send_sge)
-{
-    doca_error_t err = doca_eth_txq_cap_is_type_supported(devinfo, DOCA_ETH_TXQ_TYPE_REGULAR,
-                                                          DOCA_ETH_TXQ_DATA_PATH_TYPE_CPU);
-    if (DOCA_IS_ERROR(err)) {
-        PRINT_DOCA_ERR(hwqtx_logerr, err, "doca_eth_txq_cap_is_type_supported");
-        return false;
-    }
-
-    err = doca_eth_txq_cap_get_max_send_buf_list_len(devinfo, &max_send_sge);
-    if (DOCA_IS_ERROR(err)) {
-        PRINT_DOCA_ERR(hwqtx_logerr, err, "doca_eth_txq_cap_get_max_send_buf_list_len");
-        return false;
-    }
-
-    err = doca_eth_txq_cap_get_max_burst_size(devinfo, max_send_sge, DOCA_MAX_LSO_HEADER,
-                                              &max_burst_size);
-    if (DOCA_IS_ERROR(err)) {
-        PRINT_DOCA_ERR(hwqtx_logerr, err, "doca_eth_txq_cap_get_max_burst_size");
-        return false;
-    }
-
-    uint32_t txq_burst_size_conf = (align32pow2(safe_mce_sys().tx_num_wr));
-    if (txq_burst_size_conf > max_burst_size) {
-        // TODO: Currently we will always have this warning... tx_num_wr set to 32K.
-        hwqtx_loginfo("Decreasing BurstSize %u to capability %u.", txq_burst_size_conf,
-                      max_burst_size);
-        txq_burst_size_conf = max_burst_size;
-    }
-    max_burst_size = txq_burst_size_conf;
-
-    err = doca_eth_txq_cap_is_l3_chksum_offload_supported(devinfo);
-    if (DOCA_IS_ERROR(err)) {
-        PRINT_DOCA_ERR(hwqtx_logerr, err, "doca_eth_txq_cap_is_l3_chksum_offload_supported");
-        return false;
-    }
-
-    err = doca_eth_txq_cap_is_l4_chksum_offload_supported(devinfo);
-    if (DOCA_IS_ERROR(err)) {
-        PRINT_DOCA_ERR(hwqtx_logerr, err, "doca_eth_txq_cap_is_l4_chksum_offload_supported");
-        return false;
-    }
-
-    return true;
-}
-
-bool hw_queue_tx::prepare_doca_txq()
-{
-    doca_error_t err;
-    doca_dev *dev = m_p_ib_ctx_handler->get_doca_device();
-    doca_devinfo *devinfo = doca_dev_as_devinfo(dev);
-    uint32_t max_burst_size = 0U;
-    uint32_t max_send_sge = 0U;
-
-    if (!check_doca_caps(devinfo, max_burst_size, max_send_sge)) {
-        hwqtx_logerr("TXQ caps failed, Dev:%s", m_p_ib_ctx_handler->get_ibname().c_str());
-        return false;
-    }
-
-    doca_eth_txq *txq = nullptr;
-    err = doca_eth_txq_create(dev, max_burst_size, &txq);
-    if (DOCA_IS_ERROR(err)) {
-        PRINT_DOCA_ERR(hwqtx_logerr, err, "doca_eth_txq_create");
-        return false;
-    }
-    m_doca_txq.reset(txq);
-    m_doca_ctx_txq = doca_eth_txq_as_doca_ctx(m_doca_txq.get());
-
-    err = doca_eth_txq_set_max_send_buf_list_len(txq, max_send_sge);
-    if (DOCA_IS_ERROR(err)) {
-        PRINT_DOCA_ERR(hwqtx_logerr, err, "doca_eth_txq_set_max_send_buf_list_len");
-        return false;
-    }
-
-    /*  Issues with mss configuration per txq:
-        1. LSO will not work if remote side decided to reduce MSS.
-        2. We reduce mss to allow LSO for both IPv4 and IPv6. It means IPv4 packets will be smaller
-       by 20 bytes, so the performance will be slightly worse than what we can get with bigger
-       packets.
-    */
-    // err = doca_eth_txq_set_mss(m_doca_txq.get(), m_p_ring->get_mtu() - IPV6_HLEN - TCP_HLEN);
-    err = doca_eth_txq_set_mss(m_doca_txq.get(), m_p_ring->get_mtu() - 20 - TCP_HLEN);
-    if (DOCA_IS_ERROR(err)) {
-        PRINT_DOCA_ERR(hwqtx_logerr, err, "doca_eth_txq_set_mss");
-        return false;
-    }
-
-    err = doca_eth_txq_set_max_lso_header_size(m_doca_txq.get(), DOCA_MAX_LSO_HEADER);
-    if (DOCA_IS_ERROR(err)) {
-        PRINT_DOCA_ERR(hwqtx_logerr, err, "doca_eth_txq_set_max_lso_header_size");
-        return false;
-    }
-
-    err = doca_ctx_set_user_data(m_doca_ctx_txq, {.ptr = this});
-    if (DOCA_IS_ERROR(err)) {
-        PRINT_DOCA_ERR(hwqtx_logerr, err, "doca_ctx_set_user_data ctx/hw_queue_tx: %p,%p",
-                       m_doca_ctx_txq, this);
-        return false;
-    }
-
-    err = doca_eth_txq_set_type(m_doca_txq.get(), DOCA_ETH_TXQ_TYPE_REGULAR);
-    if (DOCA_IS_ERROR(err)) {
-        PRINT_DOCA_ERR(hwqtx_logerr, err, "doca_eth_txq_get_type_supported");
-        return false;
-    }
-
-    err = doca_eth_txq_task_send_set_conf(m_doca_txq.get(), tx_task_completion_cb, tx_task_error_cb,
-                                          max_burst_size);
-    if (DOCA_IS_ERROR(err)) {
-        PRINT_DOCA_ERR(hwqtx_logerr, err, "doca_eth_txq_task_send_set_conf txq: %p max-tasks: %u",
-                       txq, max_burst_size);
-        return false;
-    }
-
-    err = doca_eth_txq_task_lso_send_set_conf(m_doca_txq.get(), tx_task_lso_completion_cb,
-                                              tx_task_lso_error_cb, max_burst_size);
-    if (DOCA_IS_ERROR(err)) {
-        PRINT_DOCA_ERR(hwqtx_logerr, err,
-                       "doca_eth_txq_task_lso_send_set_conf txq: %p max-tasks: %u",
-                       m_doca_txq.get(), max_burst_size);
-        return false;
-    }
-
-    m_task_list_count = max_burst_size;
-
-    err = doca_eth_txq_set_l3_chksum_offload(txq, DOCA_CHECKSUM_HW_L3_ENABLE);
-    if (DOCA_IS_ERROR(err)) {
-        PRINT_DOCA_ERR(hwqtx_logerr, err, "doca_eth_txq_set_l3_chksum_offload txq: %p",
-                       m_doca_txq.get());
-        return false;
-    }
-
-    err = doca_eth_txq_set_l4_chksum_offload(txq, DOCA_CHECKSUM_HW_L4_ENABLE);
-    if (DOCA_IS_ERROR(err)) {
-        PRINT_DOCA_ERR(hwqtx_logerr, err, "doca_eth_txq_set_l4_chksum_offload txq: %p",
-                       m_doca_txq.get());
-        return false;
-    }
-
-    doca_pe *pe;
-    err = doca_pe_create(&pe);
-    if (DOCA_IS_ERROR(err)) {
-        PRINT_DOCA_ERR(hwqtx_logerr, err, "doca_pe_create");
-        return false;
-    }
-    m_doca_pe.reset(pe);
-
-    err = doca_pe_connect_ctx(pe, m_doca_ctx_txq);
-    if (DOCA_IS_ERROR(err)) {
-        PRINT_DOCA_ERR(hwqtx_logerr, err, "doca_pe_connect_ctx pe/ctx/txq: %p,%p,%p", pe,
-                       m_doca_ctx_txq, m_doca_txq.get());
-        return false;
-    }
-
-    err = doca_pe_set_event_mode(pe, DOCA_PE_EVENT_MODE_PROGRESS_ALL);
-    if (DOCA_IS_ERROR(err)) {
-        PRINT_DOCA_ERR(hwqtx_logerr, err, "doca_pe_set_event_mode pe: %p", pe);
-        return false;
-    }
-
-    err = doca_pe_get_notification_handle(pe, &m_notification_handle);
-    if (DOCA_IS_ERROR(err)) {
-        PRINT_DOCA_ERR(hwqtx_logerr, err, "doca_pe_get_notification_handle");
-        return false;
-    }
-
-#if defined(DEFINED_NGINX) || defined(DEFINED_ENVOY)
-    if (safe_mce_sys().app.distribute_cq_interrupts) {
-        uint32_t num_comp_vectors = 0;
-        err = doca_ctx_cap_get_num_completion_vectors(devinfo, &num_comp_vectors);
-        if (DOCA_IS_ERROR(err)) {
-            PRINT_DOCA_ERR(hwqtx_logerr, err, "doca_ctx_cap_get_num_completion_vectors devinfo: %p",
-                           devinfo);
-            return false;
-        }
-
-        // fetching once - as this operation requires locking
-        const int worker_id = g_p_app->get_worker_id();
-        if (likely(worker_id >= 0)) {
-            const uint32_t comp_vector = worker_id % num_comp_vectors;
-            hwqtx_logdbg("Setting PE completion affinity: %" PRIu32 ", pid: %d", comp_vector,
-                         getpid());
-            err = doca_ctx_set_completion_vector(m_doca_ctx_txq, comp_vector);
-            if (DOCA_IS_ERROR(err)) {
-                PRINT_DOCA_ERR(hwqtx_logerr, err,
-                               "doca_ctx_set_completion_vector ctx/comp_vector: %p,%" PRIu32,
-                               m_doca_ctx_txq, comp_vector);
-            }
-        }
-    }
-#endif
-
-    doca_buf_inventory *inventory = nullptr;
-    err = doca_buf_inventory_create(max_burst_size, &inventory);
-    if (DOCA_IS_ERROR(err)) {
-        PRINT_DOCA_ERR(hwqtx_logerr, err, "doca_buf_inventory_create");
-        return false;
-    }
-    m_doca_inventory.reset(inventory);
-
-    err = doca_buf_inventory_start(inventory);
-    if (DOCA_IS_ERROR(err)) {
-        PRINT_DOCA_ERR(hwqtx_logerr, err, "doca_buf_inventory_start");
-        return false;
-    }
-
-    hwqtx_loginfo("Creating DOCA TXQ MaxBurstSize: %u, Dev:%s", max_burst_size,
-                  m_p_ib_ctx_handler->get_ibname().c_str());
-    return true;
-}
-
-void hw_queue_tx::start_doca_txq()
-{
-    doca_error_t err = doca_ctx_start(m_doca_ctx_txq);
-    if (DOCA_IS_ERROR(err)) {
-        PRINT_DOCA_ERR(hwqtx_logerr, err, "doca_ctx_start(TXQ). TXQ:%p", m_doca_txq.get());
-        return;
-    }
-
-    hwqtx_loginfo("DOCA TXQ started, ctx: %p", m_doca_ctx_txq);
-}
-
-void hw_queue_tx::stop_doca_txq()
-{
-    hwqtx_logdbg("Stopping DOCA TXQ: %p", m_doca_txq.get());
-
-    doca_error_t err = doca_ctx_stop(m_doca_ctx_txq);
-    if (DOCA_ERROR_IN_PROGRESS == err) {
-        doca_ctx_states ctx_state = DOCA_CTX_STATE_STOPPING; // Just to enter the while loop.
-        while (DOCA_CTX_STATE_IDLE != ctx_state) {
-            if (!doca_pe_progress(m_doca_pe.get())) {
-                err = doca_ctx_get_state(m_doca_ctx_txq, &ctx_state);
-                if (err != DOCA_SUCCESS) {
-                    PRINT_DOCA_ERR(hwqtx_logerr, err,
-                                   "Error flushing DOCA TXQ (doca_ctx_get_state). TXQ:%p",
-                                   m_doca_txq.get());
-                    break;
-                }
-            }
-        }
-    } else if (DOCA_IS_ERROR(err)) {
-        PRINT_DOCA_ERR(hwqtx_logerr, err, "doca_ctx_stop(TXQ). TXQ:%p", m_doca_txq.get());
-    }
 }
 
 int hw_queue_tx::configure(const slave_data_t *slave,
@@ -531,7 +265,7 @@ int hw_queue_tx::configure(const slave_data_t *slave,
     m_mlx5_qp.cap.max_inline_data =
         std::min(tmp_ibv_qp_attr.cap.max_inline_data, m_mlx5_qp.cap.max_inline_data);
 
-    hwqtx_logdbg("Used QP (num=%d) wre: tx = %d sge: tx = %d inline: %d", m_mlx5_qp.qp->qp_num,
+    hwqtx_logdbg("Succeeded to create TX QP (num=%d) wre: tx = %d sge: tx = %d inline: %d", m_mlx5_qp.qp->qp_num,
                  m_mlx5_qp.cap.max_send_wr, m_mlx5_qp.cap.max_send_sge,
                  m_mlx5_qp.cap.max_inline_data);
 
@@ -559,50 +293,6 @@ int hw_queue_tx::configure(const slave_data_t *slave,
 #endif /* DEFINED_ROCE_LAG */
     NOT_IN_USE(slave);
     return 0;
-}
-
-void hw_queue_tx::up()
-{
-    if (safe_mce_sys().doca_tx) {
-        start_doca_txq();
-    }
-
-    init_queue();
-
-    // Add buffers
-    hwqtx_logdbg("QP current state: %d", priv_ibv_query_qp_state(m_mlx5_qp.qp));
-
-    m_p_cq_mgr_tx->add_qp_tx(this);
-
-    release_tx_buffers();
-
-    modify_queue_to_ready_state();
-
-    init_device_memory();
-}
-
-void hw_queue_tx::down()
-{
-    if (safe_mce_sys().doca_tx) {
-        stop_doca_txq();
-    }
-
-    if (m_dm_enabled) {
-        m_dm_mgr.release_resources();
-    }
-
-    hwqtx_logdbg("QP current state: %d", priv_ibv_query_qp_state(m_mlx5_qp.qp));
-    modify_queue_to_error_state();
-
-    // free buffers from current active resource iterator
-    trigger_completion_for_all_sent_packets();
-
-    // let the QP drain all wqe's to flushed cqe's now that we moved
-    // it to error state and post_sent final trigger for completion
-    usleep(1000);
-
-    release_tx_buffers();
-    m_p_cq_mgr_tx->del_qp_tx(this);
 }
 
 void hw_queue_tx::release_tx_buffers()
@@ -744,8 +434,8 @@ void hw_queue_tx::init_queue()
         m_sq_wqe_prop_last = nullptr;
     }
 
-    hwqtx_logfunc("m_tx_num_wr=%d max_inline_data: %d m_sq_wqe_idx_to_prop=%p", m_tx_num_wr,
-                  get_max_inline_data(), m_sq_wqe_idx_to_prop);
+    hwqtx_loginfo("Succeeded to init TX QP m_tx_num_wr=%d max_inline_data: %d m_max_sge=%u", m_tx_num_wr,
+                  get_max_inline_data(), m_mlx5_qp.cap.max_send_sge);
 
     memset((void *)(uintptr_t)m_sq_wqe_hot, 0, sizeof(struct mlx5_eth_wqe));
     m_sq_wqe_hot->ctrl.data[0] = htonl(MLX5_OPCODE_SEND);
@@ -1103,11 +793,6 @@ void hw_queue_tx::store_current_wqe_prop(mem_buf_desc_t *buf, unsigned credits, 
 void hw_queue_tx::send_to_wire(xlio_ibv_send_wr *p_send_wqe, xlio_wr_tx_packet_attr attr,
                                bool request_comp, xlio_tis *tis, unsigned credits)
 {
-    if (safe_mce_sys().doca_tx) {
-        hwqtx_logerr("Should use doca_tx path");
-        return;
-    }
-
     struct xlio_mlx5_wqe_ctrl_seg *ctrl = nullptr;
     struct mlx5_wqe_eth_seg *eseg = nullptr;
     uint32_t tisn = tis ? tis->get_tisn() : 0;
@@ -1307,55 +992,6 @@ void hw_queue_tx::tls_context_resync_tx(const xlio_tls_info *info, xlio_tis *tis
     tls_post_progress_params_wqe(tis, tisn, 0, skip_static, true);
     m_b_fence_needed = true;
 }
-
-#ifdef DEFINED_DPCP_PATH_RX
-int hw_queue_tx::tls_context_setup_rx(xlio_tir *tir, const xlio_tls_info *info,
-                                      uint32_t next_record_tcp_sn, xlio_comp_cb_t callback,
-                                      void *callback_arg)
-{
-    uint32_t tirn;
-    dpcp::tls_dek *_dek;
-    dpcp::status status;
-    dpcp::adapter *adapter = m_p_ib_ctx_handler->get_dpcp_adapter();
-    struct dpcp::dek_attr dek_attr;
-
-    memset(&dek_attr, 0, sizeof(dek_attr));
-    dek_attr.key_blob = (void *)info->key;
-    dek_attr.key_blob_size = info->key_len;
-    dek_attr.key_size = info->key_len;
-    dek_attr.pd_id = adapter->get_pd();
-    status = adapter->create_tls_dek(dek_attr, _dek);
-    if (unlikely(status != dpcp::DPCP_OK)) {
-        hwqtx_logerr("Failed to create DEK, status: %d", status);
-        return -1;
-    }
-    tir->assign_dek(_dek);
-    tir->assign_callback(callback, callback_arg);
-    tirn = tir->get_tirn();
-
-    tls_post_static_params_wqe(NULL, info, tirn, _dek->get_key_id(), 0, false, false);
-    tls_post_progress_params_wqe(tir, tirn, next_record_tcp_sn, false, false);
-
-    assert(!tir->m_released);
-
-    return 0;
-}
-
-void hw_queue_tx::tls_resync_rx(xlio_tir *tir, const xlio_tls_info *info, uint32_t hw_resync_tcp_sn)
-{
-    tls_post_static_params_wqe(tir, info, tir->get_tirn(), tir->get_dek_id(), hw_resync_tcp_sn,
-                               false, false);
-}
-
-void hw_queue_tx::tls_get_progress_params_rx(xlio_tir *tir, void *buf, uint32_t lkey)
-{
-    /* Address must be aligned by 64. */
-    assert((uintptr_t)buf == ((uintptr_t)buf >> 6U << 6U));
-
-    tls_get_progress_params_wqe(tir, tir->get_tirn(), buf, lkey);
-}
-
-#endif // DEFINED_DPCP_PATH_RX
 
 inline void hw_queue_tx::tls_fill_static_params_wqe(struct mlx5_wqe_tls_static_params_seg *params,
                                                     const struct xlio_tls_info *info,
@@ -1646,10 +1282,6 @@ void hw_queue_tx::post_dump_wqe(xlio_tis *tis, void *addr, uint32_t len, uint32_
 // So we can post_send anything we want :)
 void hw_queue_tx::trigger_completion_for_all_sent_packets()
 {
-    if (safe_mce_sys().doca_tx) {
-        return;
-    }
-
     hwqtx_logfunc("unsignaled count=%d", m_n_unsignaled_count);
 
     if (!is_signal_requested_for_last_wqe()) {
@@ -1723,6 +1355,42 @@ void hw_queue_tx::reset_inflight_zc_buffers_ctx(void *ctx)
     }
 }
 
+void hw_queue_tx::up()
+{
+    init_queue();
+
+    // Add buffers
+    hwqtx_logdbg("QP current state: %d", priv_ibv_query_qp_state(m_mlx5_qp.qp));
+
+    m_p_cq_mgr_tx->add_qp_tx(this);
+
+    release_tx_buffers();
+
+    modify_queue_to_ready_state();
+
+    init_device_memory();
+}
+
+void hw_queue_tx::down()
+{
+    if (m_dm_enabled) {
+        m_dm_mgr.release_resources();
+    }
+
+    hwqtx_logdbg("QP current state: %d", priv_ibv_query_qp_state(m_mlx5_qp.qp));
+    modify_queue_to_error_state();
+
+    // free buffers from current active resource iterator
+    trigger_completion_for_all_sent_packets();
+
+    // let the QP drain all wqe's to flushed cqe's now that we moved
+    // it to error state and post_sent final trigger for completion
+    usleep(1000);
+
+    release_tx_buffers();
+    m_p_cq_mgr_tx->del_qp_tx(this);
+}
+
 uint32_t hw_queue_tx::is_ratelimit_change(struct xlio_rate_limit_t &rate_limit)
 {
     uint32_t rl_changes = 0;
@@ -1754,365 +1422,53 @@ int hw_queue_tx::modify_qp_ratelimit(struct xlio_rate_limit_t &rate_limit, uint3
     return 0;
 }
 
-void hw_queue_tx::destory_doca_txq(doca_eth_txq *txq)
+#ifdef DEFINED_DPCP_PATH_RX
+int hw_queue_tx::tls_context_setup_rx(xlio_tir *tir, const xlio_tls_info *info,
+                                      uint32_t next_record_tcp_sn, xlio_comp_cb_t callback,
+                                      void *callback_arg)
 {
-    doca_error_t err = doca_eth_txq_destroy(txq);
-    if (DOCA_IS_ERROR(err)) {
-        PRINT_DOCA_ERR(__log_err, err, "doca_eth_txq_destroy txq: %p", txq);
+    uint32_t tirn;
+    dpcp::tls_dek *_dek;
+    dpcp::status status;
+    dpcp::adapter *adapter = m_p_ib_ctx_handler->get_dpcp_adapter();
+    struct dpcp::dek_attr dek_attr;
+
+    memset(&dek_attr, 0, sizeof(dek_attr));
+    dek_attr.key_blob = (void *)info->key;
+    dek_attr.key_blob_size = info->key_len;
+    dek_attr.key_size = info->key_len;
+    dek_attr.pd_id = adapter->get_pd();
+    status = adapter->create_tls_dek(dek_attr, _dek);
+    if (unlikely(status != dpcp::DPCP_OK)) {
+        hwqtx_logerr("Failed to create DEK, status: %d", status);
+        return -1;
     }
+    tir->assign_dek(_dek);
+    tir->assign_callback(callback, callback_arg);
+    tirn = tir->get_tirn();
+
+    tls_post_static_params_wqe(NULL, info, tirn, _dek->get_key_id(), 0, false, false);
+    tls_post_progress_params_wqe(tir, tirn, next_record_tcp_sn, false, false);
+
+    assert(!tir->m_released);
+
+    return 0;
 }
 
-void hw_queue_tx::destory_doca_inventory(doca_buf_inventory *inv)
+void hw_queue_tx::tls_resync_rx(xlio_tir *tir, const xlio_tls_info *info, uint32_t hw_resync_tcp_sn)
 {
-    doca_error_t err = doca_buf_inventory_destroy(inv);
-    if (DOCA_IS_ERROR(err)) {
-        PRINT_DOCA_ERR(__log_err, err, "doca_buf_inventory_destroy Inv: %p", inv);
-    }
+    tls_post_static_params_wqe(tir, info, tir->get_tirn(), tir->get_dek_id(), hw_resync_tcp_sn,
+                               false, false);
 }
 
-void hw_queue_tx::destory_doca_pe(doca_pe *pe)
+void hw_queue_tx::tls_get_progress_params_rx(xlio_tir *tir, void *buf, uint32_t lkey)
 {
-    doca_error_t err = doca_pe_destroy(pe);
-    if (DOCA_IS_ERROR(err)) {
-        PRINT_DOCA_ERR(__log_err, err, "doca_pe_destroy PE: %p", pe);
-    }
+    /* Address must be aligned by 64. */
+    assert((uintptr_t)buf == ((uintptr_t)buf >> 6U << 6U));
+
+    tls_get_progress_params_wqe(tir, tir->get_tirn(), buf, lkey);
 }
 
-void hw_queue_tx::tx_task_completion_cb(doca_eth_txq_task_send *task_send, doca_data task_user_data,
-                                        doca_data ctx_user_data)
-{
-    mem_buf_desc_t *mem_buf = reinterpret_cast<mem_buf_desc_t *>(task_user_data.ptr);
-    hw_queue_tx *hw_tx = reinterpret_cast<hw_queue_tx *>(ctx_user_data.ptr);
+#endif // DEFINED_DPCP_PATH_RX
 
-    hw_tx->return_doca_task(task_send);
-    hw_tx->m_p_ring->put_tx_buffer_helper(mem_buf);
-}
-
-void hw_queue_tx::tx_task_lso_completion_cb(doca_eth_txq_task_lso_send *lso_task,
-                                            doca_data task_user_data, doca_data ctx_user_data)
-{
-    doca_lso_metadata *lso_metadata = reinterpret_cast<doca_lso_metadata *>(task_user_data.ptr);
-    hw_queue_tx *hw_tx = reinterpret_cast<hw_queue_tx *>(ctx_user_data.ptr);
-
-    hw_tx->return_doca_lso_task(lso_task);
-    hw_tx->m_p_ring->put_tx_buffer_helper(lso_metadata->buff);
-    hw_tx->put_lso_metadata(lso_metadata);
-}
-
-void hw_queue_tx::tx_task_error_cb(doca_eth_txq_task_send *task_send, doca_data task_user_data,
-                                   doca_data ctx_user_data)
-{
-    mem_buf_desc_t *mem_buf = reinterpret_cast<mem_buf_desc_t *>(task_user_data.ptr);
-    hw_queue_tx *hw_tx = reinterpret_cast<hw_queue_tx *>(ctx_user_data.ptr);
-    doca_ctx_states ctx_state = DOCA_CTX_STATE_STOPPING;
-    doca_error_t rc_state = doca_ctx_get_state(hw_tx->m_doca_ctx_txq, &ctx_state);
-    ctx_state = ((ctx_state != DOCA_CTX_STATE_IDLE) ? ctx_state : DOCA_CTX_STATE_STOPPING);
-    if (rc_state != DOCA_SUCCESS || ctx_state != DOCA_CTX_STATE_STOPPING) {
-        PRINT_DOCA_ERR(__log_err,
-                       doca_task_get_status(doca_eth_txq_task_send_as_doca_task(task_send)),
-                       "TX Task Error");
-        // TODO DOCA: Add statistics for errors
-    }
-
-    hw_tx->return_doca_task(task_send);
-    hw_tx->m_p_ring->put_tx_buffer_helper(mem_buf);
-}
-
-void hw_queue_tx::tx_task_lso_error_cb(doca_eth_txq_task_lso_send *lso_task,
-                                       doca_data task_user_data, doca_data ctx_user_data)
-{
-    doca_lso_metadata *lso_metadata = reinterpret_cast<doca_lso_metadata *>(task_user_data.ptr);
-    hw_queue_tx *hw_tx = reinterpret_cast<hw_queue_tx *>(ctx_user_data.ptr);
-    doca_ctx_states ctx_state = DOCA_CTX_STATE_STOPPING;
-    doca_error_t rc_state = doca_ctx_get_state(hw_tx->m_doca_ctx_txq, &ctx_state);
-    ctx_state = ((ctx_state != DOCA_CTX_STATE_IDLE) ? ctx_state : DOCA_CTX_STATE_STOPPING);
-    if (rc_state != DOCA_SUCCESS || ctx_state != DOCA_CTX_STATE_STOPPING) {
-        PRINT_DOCA_ERR(__log_err,
-                       doca_task_get_status(doca_eth_txq_task_lso_send_as_doca_task(lso_task)),
-                       "TX Task Error");
-        // TODO DOCA: Add statistics for errors
-    }
-
-    __log_func("tx_task_lso_error_cb, lso_task: %p, rc_state: %d, ctx_state: %d", lso_task,
-               rc_state, ctx_state);
-
-    hw_tx->return_doca_lso_task(lso_task);
-    hw_tx->m_p_ring->put_tx_buffer_helper(lso_metadata->buff);
-    hw_tx->put_lso_metadata(lso_metadata);
-}
-
-void hw_queue_tx::return_doca_task(doca_eth_txq_task_send *task_send)
-{
-    doca_buf *buf = nullptr;
-    doca_error_t err = doca_eth_txq_task_send_get_pkt(task_send, &buf);
-    if (unlikely(DOCA_IS_ERROR(err))) {
-        PRINT_DOCA_ERR(hwqtx_logerr, err, "doca_eth_txq_task_send_get_pkt");
-    } else {
-        return_doca_buf(buf);
-    }
-
-    doca_task_free(doca_eth_txq_task_send_as_doca_task(task_send));
-}
-
-void hw_queue_tx::return_doca_lso_task(doca_eth_txq_task_lso_send *lso_task)
-{
-    doca_buf *buf = nullptr;
-    doca_error_t err = doca_eth_txq_task_lso_send_get_pkt_payload(lso_task, &buf);
-    if (unlikely(DOCA_IS_ERROR(err))) {
-        PRINT_DOCA_ERR(hwqtx_logerr, err, "doca_eth_txq_task_lso_send_get_pkt_payload");
-    } else {
-        return_doca_buf(buf);
-    }
-
-    doca_task_free(doca_eth_txq_task_lso_send_as_doca_task(lso_task));
-}
-
-void hw_queue_tx::return_doca_buf(doca_buf *buf)
-{
-    doca_error_t rc_state = doca_buf_dec_refcount(buf, nullptr);
-    if (unlikely(rc_state != DOCA_SUCCESS)) {
-        PRINT_DOCA_ERR(hwqtx_logerr, rc_state, "doca_buf_dec_refcount");
-    }
-}
-
-bool hw_queue_tx::expand_doca_inventory()
-{
-    doca_error_t rc = doca_buf_inventory_expand(m_doca_inventory.get(), DOCA_EXPAND_BATCH_SIZE);
-    if (DOCA_IS_ERROR(rc)) {
-        PRINT_DOCA_ERR(hwqtx_logerr, rc, "doca_buf_inventory_expand");
-        return false;
-    }
-    return true;
-}
-
-bool hw_queue_tx::expand_doca_task_pool(bool is_lso)
-{
-    if (m_task_list_count >= safe_mce_sys().tx_queue_max_elements) {
-        hwqtx_logfunc("Silent packet drop, can't expand task pool");
-        return false;
-    }
-
-    const uint16_t expand_batch_size =
-        (m_task_list_count + DOCA_EXPAND_BATCH_SIZE) >= safe_mce_sys().tx_queue_max_elements
-        ? (m_task_list_count + DOCA_EXPAND_BATCH_SIZE) % DOCA_EXPAND_BATCH_SIZE
-        : (DOCA_EXPAND_BATCH_SIZE);
-
-    doca_error_t rc;
-    doca_eth_txq *txq = m_doca_txq.get();
-    if (is_lso) {
-        rc = doca_eth_txq_task_lso_send_num_expand(txq, expand_batch_size);
-    } else {
-        rc = doca_eth_txq_task_send_num_expand(txq, expand_batch_size);
-    }
-
-    if (DOCA_IS_ERROR(rc)) {
-        PRINT_DOCA_ERR(hwqtx_logerr, rc, "DOCA expand task pool (LSO=%d) failed", is_lso);
-        return false;
-    }
-
-    m_task_list_count += expand_batch_size;
-    return true;
-}
-
-doca_lso_metadata *hw_queue_tx::get_lso_metadata()
-{
-    if (unlikely(!m_p_doca_lso_metadata_list)) {
-        m_p_doca_lso_metadata_list = g_lso_metadata_pool->get_objs(safe_mce_sys().lso_pool_batch);
-        if (unlikely(!m_p_doca_lso_metadata_list)) {
-            return nullptr;
-        }
-    }
-    doca_lso_metadata *ret = m_p_doca_lso_metadata_list;
-    m_p_doca_lso_metadata_list = m_p_doca_lso_metadata_list->next;
-    return ret;
-}
-
-void hw_queue_tx::put_lso_metadata(doca_lso_metadata *lso_metadata)
-{
-    lso_metadata->next = m_p_doca_lso_metadata_list;
-    m_p_doca_lso_metadata_list = lso_metadata;
-}
-
-bool hw_queue_tx::request_notification()
-{
-    if (likely(!m_notification_armed)) {
-        doca_error_t rc = doca_pe_request_notification(m_doca_pe.get());
-        if (unlikely(DOCA_IS_ERROR(rc))) {
-            PRINT_DOCA_ERR(hwqtx_logerr, rc, "doca_pe_request_notification");
-            return false;
-        }
-
-        ++m_hwq_tx_stats.n_tx_interrupt_requests;
-    }
-
-    m_notification_armed = true;
-    return true;
-}
-
-void hw_queue_tx::clear_notification()
-{
-    if (m_notification_armed) {
-        m_notification_armed = false;
-        doca_error_t rc = doca_pe_clear_notification(m_doca_pe.get(), m_notification_handle);
-        if (unlikely(DOCA_IS_ERROR(rc))) {
-            PRINT_DOCA_ERR(hwqtx_logerr, rc, "doca_pe_clear_notification");
-        } else {
-            ++m_hwq_tx_stats.n_tx_interrupt_received;
-        }
-    } else {
-        hwqtx_logwarn("Clear notification attempt on unarmed PE. hw_queue_tx: %p", this);
-    }
-}
-
-/*
-    1. doca_buf_inventory_buf_get_by_data
-    2. doca task allocation
-    3. doca_task_submit
-    In case step 1 and 2 fail - we try to expand inventory/task pool and retry.
-*/
-uint32_t hw_queue_tx::send_doca_single(void *ptr, uint32_t len, mem_buf_desc_t *user_data)
-{
-    doca_eth_txq_task_send *task = nullptr;
-    doca_buf *tx_doca_buf = nullptr;
-
-get_buf:
-    doca_error_t rc = doca_buf_inventory_buf_get_by_data(m_doca_inventory.get(), m_doca_mmap, ptr,
-                                                         len, &tx_doca_buf);
-    if (DOCA_IS_ERROR(rc)) {
-        if ((DOCA_ERROR_NO_MEMORY == rc) && expand_doca_inventory()) {
-            goto get_buf;
-        }
-        PRINT_DOCA_ERR(hwqtx_loginfo, rc, "doca_buf_inventory_buf_get_by_data");
-        return 0;
-    }
-
-get_task:
-    rc = doca_eth_txq_task_send_allocate_init(m_doca_txq.get(), tx_doca_buf, {.ptr = user_data},
-                                              &task);
-    if (DOCA_IS_ERROR(rc)) {
-        if ((DOCA_ERROR_NO_MEMORY == rc) && expand_doca_task_pool(false)) {
-            goto get_task;
-        }
-        return_doca_buf(tx_doca_buf);
-        PRINT_DOCA_ERR(hwqtx_loginfo, rc, "doca_eth_txq_task_send_allocate_init");
-        return 0;
-    }
-
-    rc = doca_task_submit(doca_eth_txq_task_send_as_doca_task(task));
-    if (DOCA_IS_ERROR(rc)) {
-        return_doca_task(task);
-        PRINT_DOCA_ERR(hwqtx_logerr, rc, "doca_eth_txq_task_send_as_doca_task");
-        return 0;
-    }
-
-    ++m_hwq_tx_stats.n_tx_pkt_count;
-    m_hwq_tx_stats.n_tx_byte_count += len;
-    return len;
-}
-
-uint32_t hw_queue_tx::send_doca_lso(struct iovec &h, struct pbuf *p, uint16_t mss, bool is_zerocopy)
-{
-    struct doca_eth_txq_task_lso_send *task = nullptr;
-    doca_buf *tx_doca_buf = nullptr;
-
-    doca_lso_metadata *lso_metadata = get_lso_metadata();
-    if (!lso_metadata) {
-        hwqtx_logwarn("Couldn't get LSO metadata object.");
-        return 0;
-    }
-
-    struct doca_mmap *mmap = (PBUF_DESC_MDESC == p->desc.attr)
-        ? reinterpret_cast<mapping_t *>(p->desc.mdesc)->get_doca_mmap()
-        : m_doca_mmap;
-
-    /*  In case of non-zc - p->payload points to TCP header. In LSO we should split between headers
-     * and payload, therefore we move payload pointer after TCP header and reduce length of first
-     * doca pbuf.
-     */
-    void *first_pkt = (uint8_t *)p->payload + (is_zerocopy ? 0 : TCP_HLEN);
-    uint32_t first_pkt_len = p->len - (is_zerocopy ? 0 : TCP_HLEN);
-
-get_first_buf:
-    doca_error_t rc = doca_buf_inventory_buf_get_by_data(m_doca_inventory.get(), mmap, first_pkt,
-                                                         first_pkt_len, &tx_doca_buf);
-    if (DOCA_IS_ERROR(rc)) {
-        if ((DOCA_ERROR_NO_MEMORY == rc) && expand_doca_inventory()) {
-            goto get_first_buf;
-        }
-        PRINT_DOCA_ERR(hwqtx_logerr, rc, "doca_buf_inventory_buf_get_by_data");
-        return 0;
-    }
-
-    uint32_t len_sent = h.iov_len + first_pkt_len + (p->next ? p->next->tot_len : 0);
-
-    lso_metadata->headers.addr = h.iov_base;
-    lso_metadata->headers.len = h.iov_len;
-    lso_metadata->headers.next = nullptr;
-    lso_metadata->buff = reinterpret_cast<mem_buf_desc_t *>(p);
-    p = p->next;
-
-    hwqtx_logfunc("LSO first part, len=%" PRIu32 ". LSO-Headers len=%" PRIu64, first_pkt_len,
-                  lso_metadata->headers.len);
-
-    doca_buf *prev_doca_buf = tx_doca_buf;
-    doca_buf *next_doca_buf = nullptr;
-    while (p) {
-        mmap = (PBUF_DESC_MDESC == p->desc.attr)
-            ? reinterpret_cast<mapping_t *>(p->desc.mdesc)->get_doca_mmap()
-            : m_doca_mmap;
-        rc = doca_buf_inventory_buf_get_by_data(m_doca_inventory.get(), mmap, p->payload, p->len,
-                                                &next_doca_buf);
-        if (DOCA_IS_ERROR(rc)) {
-            if ((DOCA_ERROR_NO_MEMORY == rc) && expand_doca_inventory()) {
-                continue;
-            }
-            PRINT_DOCA_ERR(hwqtx_logerr, rc, "doca_buf_inventory_buf_get_by_data");
-            return_doca_buf(tx_doca_buf);
-            return 0;
-        }
-
-        hwqtx_logfunc("LSO part, len=%" PRIu32, p->len);
-
-        doca_buf_chain_list_tail(tx_doca_buf, prev_doca_buf, next_doca_buf);
-        prev_doca_buf = next_doca_buf;
-        p = p->next;
-    }
-
-get_lso_task:
-    rc = doca_eth_txq_task_lso_send_allocate_init(
-        m_doca_txq.get(), tx_doca_buf, &lso_metadata->headers, {.ptr = lso_metadata}, &task);
-    if (DOCA_IS_ERROR(rc)) {
-        if ((DOCA_ERROR_NO_MEMORY == rc) && expand_doca_task_pool(true)) {
-            goto get_lso_task;
-        }
-        PRINT_DOCA_ERR(hwqtx_logerr, rc, "doca_eth_txq_task_lso_send_allocate_init");
-        return_doca_buf(tx_doca_buf);
-        return 0;
-    }
-
-    hwqtx_logfunc("LSO Task, len_sent=%" PRIu32 ", mss=%" PRIu16, len_sent, mss);
-
-    doca_eth_txq_task_lso_send_set_mss(task, mss);
-
-    rc = doca_task_submit(doca_eth_txq_task_lso_send_as_doca_task(task));
-    if (DOCA_IS_ERROR(rc)) {
-        return_doca_lso_task(task);
-        PRINT_DOCA_ERR(hwqtx_logerr, rc, "doca_task_submit");
-        return 0;
-    }
-
-    ++m_hwq_tx_stats.n_tx_pkt_count;
-    m_hwq_tx_stats.n_tx_byte_count += len_sent;
-    ++m_hwq_tx_stats.n_tx_tso_pkt_count;
-    m_hwq_tx_stats.n_tx_tso_byte_count += len_sent;
-    return len_sent;
-}
-
-void hw_queue_tx::poll_and_process_doca_tx()
-{
-    while (doca_pe_progress(m_doca_pe.get())) {
-        ;
-    }
-
-    m_p_ring->return_to_global_pool();
-}
+#endif // DEFINED_DPCP_PATH_TX
