+++ src/core/dev/dpcp/hw_queue_tx_dpcp.h
@@ -31,43 +31,24 @@
  * SOFTWARE.
  */
 
-#ifndef HW_QUEUE_TX_H
-#define HW_QUEUE_TX_H
+#ifndef HW_QUEUE_TX_DPCP_H
+#define HW_QUEUE_TX_DPCP_H
 
+#include "config.h"
+#ifdef DEFINED_DPCP_PATH_TX
 #include <list>
 #include <vector>
 #include "dev/xlio_ti.h"
-#include "dev/cq_mgr_tx.h"
-#include "dev/dm_mgr.h"
 #include "proto/mem_buf_desc.h"
 #include "proto/xlio_lwip.h"
-#include "util/sg_array.h"
-#include <doca_eth_txq.h>
-#include <doca_pe.h>
-#include <doca_buf_inventory.h>
-#include <doca_mmap.h>
-#include <doca_ctx.h>
-#include <doca_eth_txq_cpu_data_path.h>
 #include "util/cached_obj_pool.h"
+#include "dev/dm_mgr.h"
+#include "dev/dpcp/cq_mgr_tx.h"
+#include "util/sg_array.h"
 
 #ifndef MAX_SUPPORTED_IB_INLINE_SIZE
 #define MAX_SUPPORTED_IB_INLINE_SIZE 884
 #endif
-
-/* DOCA LSO user data */
-struct doca_lso_metadata {
-    struct doca_gather_list headers;
-    union {
-        mem_buf_desc_t *buff;
-        doca_lso_metadata *next;
-    };
-};
-
-typedef cached_obj_pool<doca_lso_metadata> lso_metadata_pool;
-extern lso_metadata_pool *g_lso_metadata_pool;
-
-struct slave_data_t;
-struct xlio_tls_info;
 
 enum {
     SQ_CREDITS_UMR = 3U,
@@ -92,12 +73,16 @@ struct sq_wqe_prop {
     struct sq_wqe_prop *next;
 };
 
+struct slave_data_t;
+struct xlio_tls_info;
+class ring_simple;
+
 // @class hw_queue_tx
 // Object to manages the SQ operations. This object is used for Tx.
 // Once created it requests from the system a CQ to work with.
 class hw_queue_tx : public xlio_ti_owner {
-    friend class cq_mgr_tx;
     friend class ring_simple;
+    friend class cq_mgr_tx;
 
 public:
     hw_queue_tx(ring_simple *ring, const slave_data_t *slave,
@@ -108,11 +93,12 @@ public:
 
     void up();
     void down();
+    int modify_qp_ratelimit(struct xlio_rate_limit_t &rate_limit, uint32_t rl_changes);
+    void dm_release_data(mem_buf_desc_t *buff) { m_dm_mgr.release_data(buff); }
+    uint32_t is_ratelimit_change(struct xlio_rate_limit_t &rate_limit);
 
     void send_wqe(xlio_ibv_send_wr *p_send_wqe, xlio_wr_tx_packet_attr attr, xlio_tis *tis,
                   unsigned credits);
-
-    struct ibv_qp *get_ibv_qp() const { return m_mlx5_qp.qp; };
 
     // This function can be replaced with a parameter during ring creation.
     // chain of calls may serve as cache warm for dummy send feature.
@@ -124,25 +110,10 @@ public:
     void modify_queue_to_ready_state();
     void modify_queue_to_error_state();
     void release_tx_buffers();
-    uint32_t is_ratelimit_change(struct xlio_rate_limit_t &rate_limit);
-    int modify_qp_ratelimit(struct xlio_rate_limit_t &rate_limit, uint32_t rl_changes);
-    void dm_release_data(mem_buf_desc_t *buff) { m_dm_mgr.release_data(buff); }
-
-#ifdef DEFINED_UTLS
-#ifdef DEFINED_DPCP_PATH_RX
-    int tls_context_setup_rx(xlio_tir *tir, const xlio_tls_info *info, uint32_t next_record_tcp_sn,
-                             xlio_comp_cb_t callback, void *callback_arg);
-    void tls_resync_rx(xlio_tir *tir, const xlio_tls_info *info, uint32_t hw_resync_tcp_sn);
-    void tls_get_progress_params_rx(xlio_tir *tir, void *buf, uint32_t lkey);
-#endif // DEFINED_DPCP_PATH_RX
-    xlio_tis *tls_context_setup_tx(const xlio_tls_info *info);
-    void tls_release_tis(xlio_tis *tis);
-    void tls_context_resync_tx(const xlio_tls_info *info, xlio_tis *tis, bool skip_static);
-    void tls_tx_post_dump_wqe(xlio_tis *tis, void *addr, uint32_t len, uint32_t lkey, bool first);
-#endif /* DEFINED_UTLS */
-
-#define DPCP_TIS_FLAGS (dpcp::TIS_ATTR_TRANSPORT_DOMAIN | dpcp::TIS_ATTR_PD)
-    std::unique_ptr<xlio_tis> create_tis(uint32_t flags);
+    void post_nop_fence();
+    void post_dump_wqe(xlio_tis *tis, void *addr, uint32_t len, uint32_t lkey, bool first);
+    void reset_inflight_zc_buffers_ctx(void *ctx);
+    void credits_return(unsigned credits) { m_sq_free_credits += credits; }
 
     /* Get a memory inside a wqebb at a wqebb_num offset from the m_sq_wqe_hot and account for
      * m_sq_wqe_counter wrap-around. Use offset_in_wqebb to for the internal address. Use the
@@ -155,19 +126,6 @@ public:
                 &(*m_sq_wqes)[(m_sq_wqe_counter + wqebb_num) & (m_tx_num_wr - 1)]) +
             offset_in_wqebb);
     }
-
-    void post_nop_fence();
-    void post_dump_wqe(xlio_tis *tis, void *addr, uint32_t len, uint32_t lkey, bool first);
-
-#if defined(DEFINED_UTLS)
-    std::unique_ptr<dpcp::tls_dek> get_new_tls_dek(const void *key, uint32_t key_size_bytes);
-    std::unique_ptr<dpcp::tls_dek> get_tls_dek(const void *key, uint32_t key_size_bytes);
-    void put_tls_dek(std::unique_ptr<dpcp::tls_dek> &&dek_obj);
-#endif
-
-    void reset_inflight_zc_buffers_ctx(void *ctx);
-
-    void credits_return(unsigned credits) { m_sq_free_credits += credits; }
 
     bool credits_get(unsigned credits)
     {
@@ -218,15 +176,13 @@ public:
                 1U;
         }
     }
-    doca_notification_handle_t get_notification_handle() const { return m_notification_handle; }
 
-    uint32_t send_doca_single(void *ptr, uint32_t len, mem_buf_desc_t *user_data);
-    uint32_t send_doca_lso(struct iovec &h, struct pbuf *p, uint16_t mss, bool is_zerocopy);
-    void poll_and_process_doca_tx();
-    bool request_notification();
-    void clear_notification();
-    void put_lso_metadata(doca_lso_metadata *lso_metadata);
-
+#if defined(DEFINED_UTLS) && defined(DEFINED_DPCP_PATH_RX)
+    int tls_context_setup_rx(xlio_tir *tir, const xlio_tls_info *info, uint32_t next_record_tcp_sn,
+                             xlio_comp_cb_t callback, void *callback_arg);
+    void tls_resync_rx(xlio_tir *tir, const xlio_tls_info *info, uint32_t hw_resync_tcp_sn);
+    void tls_get_progress_params_rx(xlio_tir *tir, void *buf, uint32_t lkey);
+#endif // DEFINED_UTLS && DEFINED_DPCP_PATH_RX
 private:
     cq_mgr_tx *init_tx_cq_mgr(struct ibv_comp_channel *p_tx_comp_event_channel);
 
@@ -238,11 +194,10 @@ private:
     void update_next_wqe_hot();
     void destroy_tis_cache();
     void put_tls_tis_in_cache(xlio_tis *tis);
+    void set_unsignaled_count(void) { m_n_unsignaled_count = m_n_sysvar_tx_num_wr_to_signal - 1; }
 
     void send_to_wire(xlio_ibv_send_wr *p_send_wqe, xlio_wr_tx_packet_attr attr, bool request_comp,
                       xlio_tis *tis, unsigned credits);
-
-    void set_unsignaled_count(void) { m_n_unsignaled_count = m_n_sysvar_tx_num_wr_to_signal - 1; }
 
     bool is_completion_need() const
     {
@@ -270,20 +225,6 @@ private:
              (m_tx_num_wr + prev_i - m_sq_wqe_prop_last_signalled) % m_tx_num_wr);
     }
 
-#if defined(DEFINED_UTLS)
-    inline void tls_fill_static_params_wqe(struct mlx5_wqe_tls_static_params_seg *params,
-                                           const struct xlio_tls_info *info, uint32_t key_id,
-                                           uint32_t resync_tcp_sn);
-    inline void tls_post_static_params_wqe(xlio_ti *ti, const struct xlio_tls_info *info,
-                                           uint32_t tis_tir_number, uint32_t key_id,
-                                           uint32_t resync_tcp_sn, bool fence, bool is_tx);
-    inline void tls_fill_progress_params_wqe(struct mlx5_wqe_tls_progress_params_seg *params,
-                                             uint32_t tis_tir_number, uint32_t next_record_tcp_sn);
-    inline void tls_post_progress_params_wqe(xlio_ti *ti, uint32_t tis_tir_number,
-                                             uint32_t next_record_tcp_sn, bool fence, bool is_tx);
-    inline void tls_get_progress_params_wqe(xlio_ti *ti, uint32_t tirn, void *buf, uint32_t lkey);
-#endif /* DEFINED_UTLS */
-
     inline void store_current_wqe_prop(mem_buf_desc_t *wr_id, unsigned credits, xlio_ti *ti);
     inline int fill_wqe(xlio_ibv_send_wr *p_send_wqe);
     inline int fill_wqe_send(xlio_ibv_send_wr *pswr);
@@ -292,18 +233,14 @@ private:
                                 int max_inline_len, int inline_len);
     inline void ring_doorbell(int num_wqebb, bool skip_comp = false);
 
-    struct xlio_rate_limit_t m_rate_limit;
+private:
     xlio_ib_mlx5_qp_t m_mlx5_qp;
-    ring_simple *m_p_ring;
     cq_mgr_tx *m_p_cq_mgr_tx;
-    ib_ctx_handler *m_p_ib_ctx_handler;
     sq_wqe_prop *m_sq_wqe_idx_to_prop = nullptr;
     sq_wqe_prop *m_sq_wqe_prop_last = nullptr;
-
     struct mlx5_eth_wqe (*m_sq_wqes)[] = nullptr;
     struct mlx5_eth_wqe *m_sq_wqe_hot = nullptr;
     uint8_t *m_sq_wqes_end = nullptr;
-
     const uint32_t m_n_sysvar_tx_num_wr_to_signal;
     uint32_t m_tx_num_wr;
     unsigned m_sq_wqe_prop_last_signalled = 0U;
@@ -315,52 +252,45 @@ private:
     bool m_b_fence_needed = false;
     bool m_dm_enabled = false;
     bool m_hw_dummy_send_support = false;
-    bool m_notification_armed = false;
-    uint8_t m_doca_max_sge = 0U;
+    hw_queue_tx_stats_t m_hwq_tx_stats;
+    ring_simple *m_p_ring;
+    ib_ctx_handler *m_p_ib_ctx_handler;
     dm_mgr m_dm_mgr;
-
+    struct xlio_rate_limit_t m_rate_limit;
+    uint8_t m_doca_max_sge = 0U;
+    bool m_notification_armed = false;
     // TIS cache. Protected by ring tx lock. TODO Move to ring.
     std::vector<xlio_tis *> m_tls_tis_cache;
 
-#if defined(DEFINED_UTLS)
+#ifdef DEFINED_UTLS
+public:
+#define DPCP_TIS_FLAGS (dpcp::TIS_ATTR_TRANSPORT_DOMAIN | dpcp::TIS_ATTR_PD)
+    xlio_tis *tls_context_setup_tx(const xlio_tls_info *info);
+    void tls_release_tis(xlio_tis *tis);
+    void tls_context_resync_tx(const xlio_tls_info *info, xlio_tis *tis, bool skip_static);
+    void tls_tx_post_dump_wqe(xlio_tis *tis, void *addr, uint32_t len, uint32_t lkey, bool first);
+    std::unique_ptr<xlio_tis> create_tis(uint32_t flags);
+    std::unique_ptr<dpcp::tls_dek> get_new_tls_dek(const void *key, uint32_t key_size_bytes);
+    std::unique_ptr<dpcp::tls_dek> get_tls_dek(const void *key, uint32_t key_size_bytes);
+    void put_tls_dek(std::unique_ptr<dpcp::tls_dek> &&dek_obj);
+
+private:
+    inline void tls_fill_static_params_wqe(struct mlx5_wqe_tls_static_params_seg *params,
+                                           const struct xlio_tls_info *info, uint32_t key_id,
+                                           uint32_t resync_tcp_sn);
+    inline void tls_post_static_params_wqe(xlio_ti *ti, const struct xlio_tls_info *info,
+                                           uint32_t tis_tir_number, uint32_t key_id,
+                                           uint32_t resync_tcp_sn, bool fence, bool is_tx);
+    inline void tls_fill_progress_params_wqe(struct mlx5_wqe_tls_progress_params_seg *params,
+                                             uint32_t tis_tir_number, uint32_t next_record_tcp_sn);
+    inline void tls_post_progress_params_wqe(xlio_ti *ti, uint32_t tis_tir_number,
+                                             uint32_t next_record_tcp_sn, bool fence, bool is_tx);
+    inline void tls_get_progress_params_wqe(xlio_ti *ti, uint32_t tirn, void *buf, uint32_t lkey);
+
     std::list<std::unique_ptr<dpcp::tls_dek>> m_tls_dek_get_cache;
     std::list<std::unique_ptr<dpcp::tls_dek>> m_tls_dek_put_cache;
 #endif
-
-    static void destory_doca_txq(doca_eth_txq *txq);
-    static void destory_doca_inventory(doca_buf_inventory *inv);
-    static void destory_doca_pe(doca_pe *pe);
-    bool prepare_doca_txq();
-
-    std::unique_ptr<doca_eth_txq, decltype(&destory_doca_txq)> m_doca_txq {nullptr,
-                                                                           destory_doca_txq};
-    std::unique_ptr<doca_buf_inventory, decltype(&destory_doca_inventory)> m_doca_inventory {
-        nullptr, destory_doca_inventory};
-    std::unique_ptr<doca_pe, decltype(&destory_doca_pe)> m_doca_pe {nullptr, destory_doca_pe};
-    doca_mmap *m_doca_mmap = nullptr;
-    doca_ctx *m_doca_ctx_txq = nullptr;
-    doca_notification_handle_t m_notification_handle;
-    doca_lso_metadata *m_p_doca_lso_metadata_list = nullptr;
-    hw_queue_tx_stats_t m_hwq_tx_stats;
-    uint32_t m_task_list_count = 0;
-
-    static void tx_task_completion_cb(doca_eth_txq_task_send *task_send, doca_data task_user_data,
-                                      doca_data ctx_user_data);
-    static void tx_task_error_cb(doca_eth_txq_task_send *task_send, doca_data task_user_data,
-                                 doca_data ctx_user_data);
-    static void tx_task_lso_completion_cb(doca_eth_txq_task_lso_send *task_send,
-                                          doca_data task_user_data, doca_data ctx_user_data);
-    static void tx_task_lso_error_cb(doca_eth_txq_task_lso_send *task_send,
-                                     doca_data task_user_data, doca_data ctx_user_data);
-    void return_doca_task(doca_eth_txq_task_send *task_send);
-    void return_doca_lso_task(doca_eth_txq_task_lso_send *lso_task);
-    void return_doca_buf(doca_buf *buf);
-    bool expand_doca_inventory();
-    bool expand_doca_task_pool(bool is_lso);
-    void start_doca_txq();
-    void stop_doca_txq();
-    bool check_doca_caps(doca_devinfo *devinfo, uint32_t &max_burst_size, uint32_t &max_send_sge);
-    doca_lso_metadata *get_lso_metadata();
 };
 
-#endif // HW_QUEUE_TX_H
+#endif // DEFINED_DPCP_PATH_TX
+#endif // HW_QUEUE_TX_DPCP_H
