+++ src/core/dev/ring.h
@@ -35,7 +35,6 @@
 #define RING_H
 
 #include <memory>
-#include "ib/base/verbs_extra.h"
 #include "dev/buffer_pool.h"
 #include "dev/xlio_ti.h"
 #include "proto/flow_tuple.h"
@@ -43,6 +42,11 @@
 #include "proto/L2_address.h"
 #include "util/cached_obj_pool.h"
 #include "lwip/tcp_impl.h"
+#include "core/lwip/pbuf.h"
+
+#ifdef DEFINED_DPCP_PATH_RX_OR_TX
+#include "ib/base/verbs_extra.h"
+#endif
 
 /* Forward declarations */
 struct xlio_tls_info;
@@ -58,8 +62,6 @@ class poll_group;
 #define ring_logfunc    __log_info_func
 #define ring_logfuncall __log_info_funcall
 #define ring_logfine    __log_info_fine
-
-typedef enum { CQT_RX, CQT_TX } cq_type_t;
 
 typedef size_t ring_user_id_t;
 
@@ -83,19 +85,11 @@ public:
     virtual int mem_buf_tx_release(mem_buf_desc_t *p_mem_buf_desc_list, bool trylock = false) = 0;
     virtual void mem_buf_rx_release(mem_buf_desc_t *p_mem_buf_desc) = 0;
 
-    virtual void send_ring_buffer(ring_user_id_t id, xlio_ibv_send_wr *p_send_wqe,
-                                  xlio_wr_tx_packet_attr attr) = 0;
-    virtual int send_lwip_buffer(ring_user_id_t id, xlio_ibv_send_wr *p_send_wqe,
-                                 xlio_wr_tx_packet_attr attr, xlio_tis *tis) = 0;
-    virtual uint32_t send_doca_single(void *ptr, uint32_t len, mem_buf_desc_t *buff) = 0;
-    virtual uint32_t send_doca_lso(struct iovec &h, struct pbuf *p, uint16_t mss,
-                                   bool is_zerocopy) = 0;
-    virtual int get_num_resources() const = 0;
     virtual size_t get_rx_channels_num() const = 0;
     virtual int get_rx_channel_fd(size_t ch_idx) const = 0;
     virtual int get_tx_channel_fd() const = 0;
-    virtual bool get_hw_dummy_send_support(ring_user_id_t id, xlio_ibv_send_wr *p_send_wqe) = 0;
-    virtual bool request_notification(cq_type_t cq_type) = 0;
+    virtual bool request_notification_tx() = 0;
+    virtual bool request_notification_rx() = 0;
     virtual bool reclaim_recv_buffers(descq_t *rx_reuse) = 0;
     virtual bool reclaim_recv_buffers(mem_buf_desc_t *rx_reuse_lst) = 0;
     virtual bool reclaim_recv_buffers_no_lock(mem_buf_desc_t *) = 0;
@@ -110,38 +104,58 @@ public:
     virtual void inc_tx_retransmissions_stats(ring_user_id_t id) = 0;
     virtual bool is_member(ring_slave *rng) = 0;
     virtual bool is_active_member(ring_slave *rng, ring_user_id_t id) = 0;
+    virtual void flow_del_all_rfs_safe() = 0;
     virtual ring_user_id_t generate_id(const address_t src_mac, const address_t dst_mac,
                                        uint16_t eth_proto, uint16_t encap_proto,
                                        const ip_address &src_ip, const ip_address &dst_ip,
                                        uint16_t src_port, uint16_t dst_port) = 0;
+    virtual uint32_t get_max_payload_sz() = 0;
+    virtual uint16_t get_max_header_sz() = 0;
+    virtual bool is_tso() = 0;
+    virtual ib_ctx_handler *get_ctx(ring_user_id_t id) = 0;
     virtual int modify_ratelimit(struct xlio_rate_limit_t &rate_limit) = 0;
+    virtual bool tls_tx_supported() = 0;
+    virtual bool tls_rx_supported() = 0;
+
+    ring *get_parent() { return m_parent; }
+    ring_user_id_t generate_id() { return 0; }
+    int get_if_index() { return m_if_index; }
+    struct tcp_seg *get_tcp_segs(uint32_t num);
+    void put_tcp_segs(struct tcp_seg *seg);
+
+#ifdef DEFINED_DPCP_PATH_TX
+    virtual bool get_hw_dummy_send_support(ring_user_id_t id, xlio_ibv_send_wr *p_send_wqe) = 0;
     virtual uint32_t get_tx_user_lkey(void *addr, size_t length) = 0;
     virtual uint32_t get_max_inline_data() = 0;
     virtual uint32_t get_max_send_sge(void) = 0;
-    virtual uint32_t get_max_payload_sz(void) = 0;
-    virtual uint16_t get_max_header_sz(void) = 0;
     virtual uint32_t get_tx_lkey(ring_user_id_t id) = 0;
-    virtual bool is_tso() = 0;
-    virtual ib_ctx_handler *get_ctx(ring_user_id_t id) = 0;
+    virtual void send_ring_buffer(ring_user_id_t id, xlio_ibv_send_wr *p_send_wqe,
+                                  xlio_wr_tx_packet_attr attr) = 0;
+    virtual int send_lwip_buffer(ring_user_id_t id, xlio_ibv_send_wr *p_send_wqe,
+                                 xlio_wr_tx_packet_attr attr, xlio_tis *tis) = 0;
     virtual std::unique_ptr<xlio_tis> create_tis(uint32_t flag) const = 0;
-    virtual void reset_inflight_zc_buffers_ctx(ring_user_id_t id, void *ctx) = 0;
-    virtual void flow_del_all_rfs_safe() = 0;
     virtual void post_nop_fence() = 0;
     virtual void post_dump_wqe(xlio_tis *tis, void *addr, uint32_t len, uint32_t lkey,
                                bool first) = 0;
+    virtual void reset_inflight_zc_buffers_ctx(ring_user_id_t id, void *ctx) = 0;
     virtual bool credits_get(unsigned credits) = 0;
     virtual void credits_return(unsigned credits) = 0;
+#else // DEFINED_DPCP_PATH_TX
+    virtual uint32_t send_doca_single(void *ptr, uint32_t len, mem_buf_desc_t *buff) = 0;
+    virtual uint32_t send_doca_lso(struct iovec &h, struct pbuf *p, uint16_t mss,
+                                   bool is_zerocopy) = 0;
+#endif // DEFINED_DPCP_PATH_TX
 
-#ifdef DEFINED_UTLS
-    virtual bool tls_tx_supported() = 0;
-    virtual bool tls_rx_supported() = 0;
-    virtual xlio_tis *tls_context_setup_tx(const xlio_tls_info *info) = 0;
+#if defined(DEFINED_DPCP_PATH_TX) && defined(DEFINED_UTLS)
     virtual void tls_release_tis(xlio_tis *tis) = 0;
+    virtual xlio_tis *tls_context_setup_tx(const xlio_tls_info *info) = 0;
     virtual void tls_context_resync_tx(const xlio_tls_info *info, xlio_tis *tis,
                                        bool skip_static) = 0;
     virtual void tls_tx_post_dump_wqe(xlio_tis *tis, void *addr, uint32_t len, uint32_t lkey,
                                       bool first) = 0;
-#ifdef DEFINED_DPCP_PATH_RX
+#endif // DEFINED_DPCP_PATH_TX && DEFINED_UTLS
+
+#if defined(DEFINED_DPCP_PATH_RX_AND_TX) && defined(DEFINED_UTLS)
     virtual xlio_tir *tls_create_tir(bool cached) = 0;
     virtual int tls_context_setup_rx(xlio_tir *tir, const xlio_tls_info *info,
                                      uint32_t next_record_tcp_sn, xlio_comp_cb_t callback,
@@ -151,14 +165,7 @@ public:
                                uint32_t hw_resync_tcp_sn) = 0;
     virtual void tls_get_progress_params_rx(xlio_tir *tir, void *buf, uint32_t lkey) = 0;
     virtual void tls_release_tir(xlio_tir *tir) = 0;
-#endif // DEFINED_DPCP_PATH_RX
-#endif /* DEFINED_UTLS */
-
-    ring *get_parent() { return m_parent; }
-    ring_user_id_t generate_id() { return 0; }
-    int get_if_index() { return m_if_index; }
-    struct tcp_seg *get_tcp_segs(uint32_t num);
-    void put_tcp_segs(struct tcp_seg *seg);
+#endif // DEFINED_DPCP_PATH_RX_AND_TX && DEFINED_UTLS
 
 protected:
     void set_parent(ring *parent) { m_parent = (parent ? parent : this); }
