+++ src/core/dev/ring_simple.cpp
@@ -33,10 +33,18 @@
 
 #include <mutex>
 #include "ring_simple.h"
-
 #include "util/valgrind.h"
-#include "util/sg_array.h"
 #include "sock/fd_collection.h"
+
+#ifdef DEFINED_DPCP_PATH_TX
+#include "util/sg_array.h"
+#define ALIGN_WR_DOWN(_num_wr_) (std::max(32, ((_num_wr_) & ~(0xf))))
+#endif // DEFINED_DPCP_PATH_TX
+
+#define RING_LOCK_AND_RUN(__lock__, __func_and_params__)                                           \
+    __lock__.lock();                                                                               \
+    __func_and_params__;                                                                           \
+    __lock__.unlock();
 
 #undef MODULE_NAME
 #define MODULE_NAME "ring_simple"
@@ -44,49 +52,15 @@ DOCA_LOG_REGISTER(ring_simple);
 #undef MODULE_HDR
 #define MODULE_HDR MODULE_NAME "%d:%s() "
 
-#define ALIGN_WR_DOWN(_num_wr_) (std::max(32, ((_num_wr_) & ~(0xf))))
 #define RING_TX_BUFS_COMPENSATE 256U
-
-#define RING_LOCK_AND_RUN(__lock__, __func_and_params__)                                           \
-    __lock__.lock();                                                                               \
-    __func_and_params__;                                                                           \
-    __lock__.unlock();
-
-#define RING_TRY_LOCK_RUN_AND_UPDATE_RET(__lock__, __func_and_params__)                            \
-    if (!__lock__.trylock()) {                                                                     \
-        ret = __func_and_params__;                                                                 \
-        __lock__.unlock();                                                                         \
-    } else {                                                                                       \
-        errno = EAGAIN;                                                                            \
-    }
-
-/**/
-/** inlining functions can only help if they are implemented before their usage **/
-/**/
-
-inline void ring_simple::send_status_handler(int ret, xlio_ibv_send_wr *p_send_wqe)
-{
-    BULLSEYE_EXCLUDE_BLOCK_START
-    if (unlikely(ret)) {
-        // Error during post_send, reclaim the tx buffer
-        if (p_send_wqe) {
-            mem_buf_desc_t *p_mem_buf_desc = (mem_buf_desc_t *)(p_send_wqe->wr_id);
-            mem_buf_tx_release(p_mem_buf_desc);
-        }
-    } else {
-        // Update TX statistics
-        sg_array sga(p_send_wqe->sg_list, p_send_wqe->num_sge);
-        m_hqtx->m_hwq_tx_stats.n_tx_byte_count += sga.length();
-        ++m_hqtx->m_hwq_tx_stats.n_tx_pkt_count;
-    }
-    BULLSEYE_EXCLUDE_BLOCK_END
-}
 
 ring_simple::ring_simple(int if_index, ring *parent, bool use_locks)
     : ring_slave(if_index, parent, use_locks)
     , m_lock_ring_tx_buf_wait("ring:lock_tx_buf_wait")
-    , m_p_doca_mmap(g_buffer_pool_tx->get_doca_mmap())
     , m_gro_mgr(safe_mce_sys().gro_streams_max, MAX_GRO_BUFS)
+#ifndef DEFINED_DPCP_PATH_TX
+    , m_p_doca_mmap(g_buffer_pool_tx->get_doca_mmap())
+#endif // DEFINED_DPCP_PATH_TX
 {
     net_device_val *p_ndev = g_p_net_device_table_mgr->get_net_device_val(m_parent->get_if_index());
     BULLSEYE_EXCLUDE_BLOCK_START
@@ -109,10 +83,12 @@ ring_simple::ring_simple(int if_index, ring *parent, bool use_locks)
         ring_logpanic("m_p_ib_ctx = NULL. It can be related to wrong bonding configuration");
     }
 
+#ifdef DEFINED_DPCP_PATH_TX
     m_tx_lkey = g_buffer_pool_tx->find_lkey_by_ib_ctx_thread_safe(m_p_ib_ctx);
     if (m_tx_lkey == 0) {
         __log_info_panic("invalid lkey found %u", m_tx_lkey);
     }
+#endif // DEFINED_DPCP_PATH_TX
     BULLSEYE_EXCLUDE_BLOCK_END
 
     /* initialization basing on ndev information */
@@ -120,9 +96,7 @@ ring_simple::ring_simple(int if_index, ring *parent, bool use_locks)
 
     memset(&m_cq_moderation_info, 0, sizeof(m_cq_moderation_info));
     memset(&m_tso, 0, sizeof(m_tso));
-#ifdef DEFINED_UTLS
     memset(&m_tls, 0, sizeof(m_tls));
-#endif /* DEFINED_UTLS */
     memset(&m_lro, 0, sizeof(m_lro));
 }
 
@@ -175,10 +149,7 @@ ring_simple::~ring_simple()
     }
 #endif // DEFINED_DPCP_PATH_RX
 
-    ring_logdbg("Tx buffer poll: free count = %lu, total = %d", m_tx_pool.size() + m_zc_pool.size(),
-                m_tx_num_bufs + m_zc_num_bufs);
-    ring_logdbg("Rx buffer pool: %lu free global buffers available", m_tx_pool.size());
-
+#ifdef DEFINED_DPCP_PATH_TX
     // Release verbs resources
     if (m_p_tx_comp_event_channel) {
         IF_VERBS_FAILURE(ibv_destroy_comp_channel(m_p_tx_comp_event_channel))
@@ -189,6 +160,11 @@ ring_simple::~ring_simple()
         VALGRIND_MAKE_MEM_UNDEFINED(m_p_tx_comp_event_channel, sizeof(struct ibv_comp_channel));
         m_p_tx_comp_event_channel = nullptr;
     }
+#endif // DEFINED_DPCP_PATH_TX
+
+    ring_logdbg("Tx buffer poll: free count = %lu, total = %d", m_tx_pool.size() + m_zc_pool.size(),
+                m_tx_num_bufs + m_zc_num_bufs);
+    ring_logdbg("Rx buffer pool: %lu free global buffers available", m_tx_pool.size());
 
     /* coverity[double_unlock] TODO: RM#1049980 */
     m_lock_ring_tx.unlock();
@@ -210,6 +186,8 @@ void ring_simple::create_resources()
     const slave_data_t *p_slave = p_ndev->get_slave(get_if_index());
 
     save_l2_address(p_slave->p_L2_addr);
+
+#ifdef DEFINED_DPCP_PATH_TX
     m_p_tx_comp_event_channel = ibv_create_comp_channel(m_p_ib_ctx->get_ibv_context());
     BULLSEYE_EXCLUDE_BLOCK_START
     if (!m_p_tx_comp_event_channel) {
@@ -236,24 +214,22 @@ void ring_simple::create_resources()
         m_tx_num_wr = max_qp_wr;
     }
     ring_logdbg("ring attributes: m_tx_num_wr = %d", m_tx_num_wr);
+#endif // DEFINED_DPCP_PATH_TX
 
     /* Detect TSO capabilities */
     memset(&m_tso, 0, sizeof(m_tso));
     if ((safe_mce_sys().enable_tso == option_3::ON) ||
         ((safe_mce_sys().enable_tso == option_3::AUTO) && (1 == validate_tso(get_if_index())))) {
-        if (xlio_check_dev_attr_tso(m_p_ib_ctx->get_ibv_device_attr())) {
-            const xlio_ibv_tso_caps *caps =
-                &xlio_get_tso_caps(m_p_ib_ctx->get_ibv_device_attr_ex());
-            if (ibv_is_qpt_supported(caps->supported_qpts, IBV_QPT_RAW_PACKET)) {
-                if (caps->max_tso && (caps->max_tso > MCE_DEFAULT_MAX_TSO_SIZE)) {
-                    ring_logwarn("max_tso cap (=%u) is higher than default TSO size (=%u). "
-                                 "Increase XLIO_MAX_TSO_SIZE to get full TSO potential.",
-                                 caps->max_tso, MCE_DEFAULT_MAX_TSO_SIZE);
-                }
-                m_tso.max_payload_sz = caps->max_tso;
-                /* ETH(14) + IP(20) + TCP(20) + TCP OPTIONS(40) */
-                m_tso.max_header_sz = 94;
+        const xlio_ibv_tso_caps *caps = &xlio_get_tso_caps(m_p_ib_ctx->get_ibv_device_attr_ex());
+        if (ibv_is_qpt_supported(caps->supported_qpts, IBV_QPT_RAW_PACKET)) {
+            if (caps->max_tso && (caps->max_tso > MCE_DEFAULT_MAX_TSO_SIZE)) {
+                ring_logwarn("max_tso cap (=%u) is higher than default TSO size (=%u). "
+                             "Increase XLIO_MAX_TSO_SIZE to get full TSO potential.",
+                             caps->max_tso, MCE_DEFAULT_MAX_TSO_SIZE);
             }
+            m_tso.max_payload_sz = caps->max_tso;
+            /* ETH(14) + IP(20) + TCP(20) + TCP OPTIONS(40) */
+            m_tso.max_header_sz = 94;
         }
     }
     ring_logdbg("ring attributes: m_tso = %d", is_tso());
@@ -348,9 +324,13 @@ void ring_simple::create_resources()
     std::unique_ptr<hw_queue_rx> temp_hqrx(new hw_queue_rx(this, p_slave->p_ib_ctx, m_vlan));
 #endif // DEFINED_DPCP_PATH_RX
 
+#ifdef DEFINED_DPCP_PATH_TX
     std::unique_ptr<hw_queue_tx> temp_hqtx(
-        new hw_queue_tx(this, p_slave, m_p_tx_comp_event_channel, get_tx_num_wr()));
+        new hw_queue_tx(this, p_slave, m_p_tx_comp_event_channel, m_tx_num_wr));
     m_p_cq_mgr_tx = temp_hqtx->get_tx_cq_mgr();
+#else // DEFINED_DPCP_PATH_TX
+    std::unique_ptr<hw_queue_tx> temp_hqtx(new hw_queue_tx(this, p_slave));
+#endif // DEFINED_DPCP_PATH_TX
 
     BULLSEYE_EXCLUDE_BLOCK_START
     if (!temp_hqtx || !temp_hqrx) {
@@ -371,11 +351,15 @@ void ring_simple::create_resources()
 
     init_tx_buffers(RING_TX_BUFS_COMPENSATE);
 
+#ifdef DEFINED_DPCP_PATH_TX
     /* For RoCE LAG device income data is processed by single ring only
      * Consider using ring related slave with lag_tx_port_affinity = 1
      * even if slave is not active
      */
     if (p_slave->active || (p_slave->lag_tx_port_affinity == 1)) {
+#else // DEFINED_DPCP_PATH_TX
+    if (p_slave->active) {
+#endif // DEFINED_DPCP_PATH_TX
         start_active_queue_tx();
         start_active_queue_rx();
     }
@@ -389,44 +373,6 @@ void ring_simple::create_resources()
                                           &m_hqrx->m_hwq_rx_stats);
 
     ring_logdbg("new ring_simple() completed");
-}
-
-int ring_simple::get_tx_channel_fd() const
-{
-    if (safe_mce_sys().doca_tx) {
-        return m_hqtx->get_notification_handle();
-    }
-
-    return m_p_tx_comp_event_channel->fd;
-}
-
-bool ring_simple::request_notification(cq_type_t cq_type)
-{
-    if (likely(CQT_RX == cq_type)) {
-        std::lock_guard<decltype(m_lock_ring_rx)> lock(m_lock_ring_rx);
-#ifdef DEFINED_DPCP_PATH_RX
-        return m_p_cq_mgr_rx->request_notification();
-#else // DEFINED_DPCP_PATH_RX
-        return m_hqrx->request_notification();
-#endif // DEFINED_DPCP_PATH_RX
-    }
-
-    std::lock_guard<decltype(m_lock_ring_tx)> lock(m_lock_ring_tx);
-    return (!safe_mce_sys().doca_tx ? m_p_cq_mgr_tx->request_notification()
-                                    : m_hqtx->request_notification());
-}
-
-void ring_simple::poll_and_process_element_tx()
-{
-    std::lock_guard<decltype(m_lock_ring_tx)> lock(m_lock_ring_tx);
-    (safe_mce_sys().doca_tx ? m_hqtx->poll_and_process_doca_tx()
-                            : m_p_cq_mgr_tx->poll_and_process_element_tx());
-}
-
-void ring_simple::mem_buf_desc_return_to_owner_tx(mem_buf_desc_t *p_mem_buf_desc)
-{
-    ring_logfuncall(LOG_FUNCTION_CALL);
-    RING_LOCK_AND_RUN(m_lock_ring_tx, put_tx_buffers(p_mem_buf_desc));
 }
 
 void ring_simple::mem_buf_desc_return_single_to_owner_tx(mem_buf_desc_t *p_mem_buf_desc)
@@ -447,19 +393,43 @@ void ring_simple::mem_buf_desc_return_single_multi_ref(mem_buf_desc_t *p_mem_buf
     put_tx_single_buffer(p_mem_buf_desc);
 }
 
+bool ring_simple::request_more_tx_buffers(pbuf_type type, uint32_t count)
+{
+    ring_logfuncall("Allocating additional %d buffers for internal use", count);
+
+#ifdef DEFINED_DPCP_PATH_TX
+    int tx_lkey = m_tx_lkey;
+#else // DEFINED_DPCP_PATH_TX
+    int tx_lkey = 0; // To be removed/refactored for DOCA
+#endif // DEFINED_DPCP_PATH_TX
+
+    bool res;
+    if (type == PBUF_ZEROCOPY) {
+        res = g_buffer_pool_zc->get_buffers_thread_safe(m_zc_pool, this, count, tx_lkey);
+    } else {
+        res = g_buffer_pool_tx->get_buffers_thread_safe(m_tx_pool, this, count, tx_lkey);
+    }
+    if (!res) {
+        ring_logfunc("Out of mem_buf_desc from TX free pool for internal object pool");
+        return false;
+    }
+
+    return true;
+}
+
 mem_buf_desc_t *ring_simple::mem_buf_tx_get(ring_user_id_t id, pbuf_type type,
                                             uint32_t n_num_mem_bufs /* default = 1 */)
 {
     std::lock_guard<decltype(m_lock_ring_tx)> lock(m_lock_ring_tx);
     NOT_IN_USE(id);
-    ring_logfuncall("n_num_mem_bufs=%d", n_num_mem_bufs);
+    ring_logfuncall("n_num_mem_bufs=%u", n_num_mem_bufs);
 
     mem_buf_desc_t *head;
     descq_t &pool = type == PBUF_ZEROCOPY ? m_zc_pool : m_tx_pool;
 
     if (unlikely(pool.size() < n_num_mem_bufs)) {
         int count = std::max(RING_TX_BUFS_COMPENSATE, n_num_mem_bufs);
-        if (request_more_tx_buffers(type, count, m_tx_lkey)) {
+        if (request_more_tx_buffers(type, count)) {
             /*
              * TODO Unify request_more_tx_buffers so ring_slave
              * keeps number of buffers instead of reinventing it in
@@ -520,160 +490,31 @@ void ring_simple::mem_buf_rx_release(mem_buf_desc_t *p_mem_buf_desc)
     reclaim_recv_buffers(p_mem_buf_desc);
 }
 
-/* note that this function is inline, so keep it above the functions using it */
-inline int ring_simple::send_buffer(xlio_ibv_send_wr *p_send_wqe, xlio_wr_tx_packet_attr attr,
-                                    xlio_tis *tis)
+int ring_simple::modify_ratelimit(struct xlio_rate_limit_t &rate_limit)
 {
-    int ret = 0;
-    unsigned credits = m_hqtx->credits_calculate(p_send_wqe);
-
-    if (likely(m_hqtx->credits_get(credits)) ||
-        is_available_qp_wr(is_set(attr, XLIO_TX_PACKET_BLOCK), credits)) {
-        m_hqtx->send_wqe(p_send_wqe, attr, tis, credits);
-    } else {
-        ring_logdbg("Silent packet drop, SQ is full!");
-        ret = -1;
-        reinterpret_cast<mem_buf_desc_t *>(p_send_wqe->wr_id)->p_next_desc = nullptr;
-        ++m_p_ring_stat->n_tx_dropped_wqes;
-    }
-    return ret;
-}
-
-bool ring_simple::get_hw_dummy_send_support(ring_user_id_t id, xlio_ibv_send_wr *p_send_wqe)
-{
-    NOT_IN_USE(id);
-    NOT_IN_USE(p_send_wqe);
-
-    return m_hqtx->get_hw_dummy_send_support();
-}
-
-void ring_simple::send_ring_buffer(ring_user_id_t id, xlio_ibv_send_wr *p_send_wqe,
-                                   xlio_wr_tx_packet_attr attr)
-{
-    NOT_IN_USE(id);
-
-    if (attr & XLIO_TX_SW_L4_CSUM) {
-        compute_tx_checksum((mem_buf_desc_t *)(p_send_wqe->wr_id), attr & XLIO_TX_PACKET_L3_CSUM,
-                            attr & XLIO_TX_PACKET_L4_CSUM);
-        attr = (xlio_wr_tx_packet_attr)(attr & ~(XLIO_TX_PACKET_L4_CSUM));
+    if (!m_p_ib_ctx->is_packet_pacing_supported(rate_limit.rate)) {
+        ring_logwarn("Packet pacing is not supported for this device");
+        return -1;
     }
 
-    std::lock_guard<decltype(m_lock_ring_tx)> lock(m_lock_ring_tx);
-    int ret = send_buffer(p_send_wqe, attr, nullptr);
-    send_status_handler(ret, p_send_wqe);
-}
+    if ((rate_limit.max_burst_sz || rate_limit.typical_pkt_sz) &&
+        !m_p_ib_ctx->get_burst_capability()) {
+        ring_logwarn("Burst is not supported for this device");
+        return -1;
+    }
 
-int ring_simple::send_lwip_buffer(ring_user_id_t id, xlio_ibv_send_wr *p_send_wqe,
-                                  xlio_wr_tx_packet_attr attr, xlio_tis *tis)
-{
-    NOT_IN_USE(id);
-    std::lock_guard<decltype(m_lock_ring_tx)> lock(m_lock_ring_tx);
-    int ret = send_buffer(p_send_wqe, attr, tis);
-    send_status_handler(ret, p_send_wqe);
-    return ret;
-}
+    uint32_t rl_changes = m_hqtx->is_ratelimit_change(rate_limit);
 
-uint32_t ring_simple::send_doca_single(void *ptr, uint32_t len, mem_buf_desc_t *buff)
-{
-    std::lock_guard<decltype(m_lock_ring_tx)> lock(m_lock_ring_tx);
-    uint32_t ret = m_hqtx->send_doca_single(ptr, len, buff);
-    m_hqtx->poll_and_process_doca_tx();
-    return ret;
-}
+    if (m_up_tx && rl_changes) {
+        return m_hqtx->modify_qp_ratelimit(rate_limit, rl_changes);
+    }
 
-uint32_t ring_simple::send_doca_lso(struct iovec &h, struct pbuf *p, uint16_t mss, bool is_zerocopy)
-{
-    std::lock_guard<decltype(m_lock_ring_tx)> lock(m_lock_ring_tx);
-    uint32_t ret = m_hqtx->send_doca_lso(h, p, mss, is_zerocopy);
-    m_hqtx->poll_and_process_doca_tx();
-    return ret;
-}
-
-/*
- * called under m_lock_ring_tx lock
- */
-bool ring_simple::is_available_qp_wr(bool b_block, unsigned credits)
-{
-    bool granted;
-    int ret;
-
-    do {
-        // Try to poll once in the hope that we get space in SQ
-        m_p_cq_mgr_tx->poll_and_process_element_tx();
-        granted = m_hqtx->credits_get(credits);
-        if (granted) {
-            break;
-        }
-
-        if (b_block) {
-            // Arm & Block on cq_mgr_tx notification channel until we get space in SQ
-
-            // Only a single thread should block on next Tx cqe event, hence the dedicated lock!
-            /* coverity[double_unlock] TODO: RM#1049980 */
-            m_lock_ring_tx.unlock();
-            m_lock_ring_tx_buf_wait.lock();
-            /* coverity[double_lock] TODO: RM#1049980 */
-            m_lock_ring_tx.lock();
-
-            // TODO Resolve race window between previous polling and request_notification
-            ret = m_p_cq_mgr_tx->request_notification();
-            if (ret < 0) {
-                // this is most likely due to cq_poll_sn out of sync, need to poll_cq again
-                ring_logdbg("failed arming cq_mgr_tx (hqtx=%p, cq_mgr_tx=%p) (errno=%d %m)", m_hqtx,
-                            m_p_cq_mgr_tx, errno);
-            } else if (ret == 0) {
-                // prepare to block
-                // CQ is armed, block on the CQ's Tx event channel (fd)
-                struct pollfd poll_fd = {/*.fd=*/0, /*.events=*/POLLIN, /*.revents=*/0};
-                poll_fd.fd = get_tx_channel_fd();
-
-                // Now it is time to release the ring lock (for restart events to be handled
-                // while this thread block on CQ channel)
-                /* coverity[double_unlock] TODO: RM#1049980 */
-                m_lock_ring_tx.unlock();
-
-                ret = SYSCALL(poll, &poll_fd, 1, -1);
-                if (ret <= 0) {
-                    ring_logdbg("failed blocking on cq_mgr_tx (errno=%d %m)", errno);
-                    m_lock_ring_tx_buf_wait.unlock();
-                    /* coverity[double_lock] TODO: RM#1049980 */
-                    m_lock_ring_tx.lock();
-                    /* coverity[missing_unlock] */
-                    return false;
-                }
-                /* coverity[double_lock] TODO: RM#1049980 */
-                m_lock_ring_tx.lock();
-
-                // Find the correct cq_mgr_tx from the CQ event,
-                // It might not be the active_cq object since we have a single TX CQ comp
-                // channel for all cq_mgr_tx's
-                cq_mgr_tx *p_cq_mgr_tx =
-                    cq_mgr_tx::get_cq_mgr_from_cq_event(m_p_tx_comp_event_channel);
-                if (p_cq_mgr_tx) {
-
-                    // Allow additional CQ arming now
-                    p_cq_mgr_tx->reset_notification_armed();
-
-                    // Perform a non blocking event read, clear the fd channel
-                    p_cq_mgr_tx->poll_and_process_element_tx();
-                }
-            }
-
-            /* coverity[double_unlock] TODO: RM#1049980 */
-            m_lock_ring_tx.unlock();
-            m_lock_ring_tx_buf_wait.unlock();
-            /* coverity[double_lock] TODO: RM#1049980 */
-            m_lock_ring_tx.lock();
-        }
-    } while (b_block);
-
-    /* coverity[missing_unlock] */
-    return granted;
+    return 0;
 }
 
 void ring_simple::init_tx_buffers(uint32_t count)
 {
-    request_more_tx_buffers(PBUF_RAM, count, m_tx_lkey);
+    request_more_tx_buffers(PBUF_RAM, count);
     m_tx_num_bufs = m_tx_pool.size();
     m_p_ring_stat->n_tx_num_bufs = m_tx_num_bufs;
 }
@@ -883,74 +724,17 @@ bool ring_simple::is_up()
     return m_up_tx && m_up_rx;
 }
 
-int ring_simple::modify_ratelimit(struct xlio_rate_limit_t &rate_limit)
-{
-    if (!m_p_ib_ctx->is_packet_pacing_supported(rate_limit.rate)) {
-        ring_logwarn("Packet pacing is not supported for this device");
-        return -1;
-    }
-
-    if ((rate_limit.max_burst_sz || rate_limit.typical_pkt_sz) &&
-        !m_p_ib_ctx->get_burst_capability()) {
-        ring_logwarn("Burst is not supported for this device");
-        return -1;
-    }
-
-    uint32_t rl_changes = m_hqtx->is_ratelimit_change(rate_limit);
-
-    if (m_up_tx && rl_changes) {
-        return m_hqtx->modify_qp_ratelimit(rate_limit, rl_changes);
-    }
-
-    return 0;
-}
-
-uint32_t ring_simple::get_tx_user_lkey(void *addr, size_t length)
-{
-    uint32_t lkey;
-
-    /*
-     * Current implementation supports a ring registration cache where addr is the key.
-     *
-     * The mode is used for send zerocopy.
-     *
-     * TODO The mode doesnn't support memory deregistration.
-     */
-    auto iter = m_user_lkey_map.find(addr);
-    if (iter != m_user_lkey_map.end()) {
-        lkey = iter->second;
-    } else {
-        lkey = m_p_ib_ctx->user_mem_reg(addr, length, XLIO_IBV_ACCESS_LOCAL_WRITE);
-        if (lkey == LKEY_ERROR) {
-            ring_logerr("Can't register user memory addr %p len %lx", addr, length);
-        } else {
-            m_user_lkey_map[addr] = lkey;
-        }
-    }
-    return lkey;
-}
-
-uint32_t ring_simple::get_max_inline_data()
-{
-    return m_hqtx->get_max_inline_data();
-}
-
-uint32_t ring_simple::get_max_send_sge(void)
-{
-    return m_hqtx->get_max_send_sge();
-}
-
-uint32_t ring_simple::get_max_payload_sz(void)
+uint32_t ring_simple::get_max_payload_sz()
 {
     return m_tso.max_payload_sz;
 }
 
-uint16_t ring_simple::get_max_header_sz(void)
+uint16_t ring_simple::get_max_header_sz()
 {
     return m_tso.max_header_sz;
 }
 
-bool ring_simple::is_tso(void)
+bool ring_simple::is_tso()
 {
     return (m_tso.max_payload_sz && m_tso.max_header_sz);
 }
