+++ src/core/dev/ring_simple.h
@@ -34,13 +34,12 @@
 #ifndef RING_SIMPLE_H
 #define RING_SIMPLE_H
 
-#include "ring_slave.h"
-
 #include <mutex>
 #include <unordered_map>
-
+#include "dev/ring_slave.h"
 #include "dev/gro_mgr.h"
-#include "dev/hw_queue_tx.h"
+#include "dev/dpcp/hw_queue_tx_dpcp.h"
+#include "dev/doca/hw_queue_tx_doca.h"
 #include "dev/dpcp/hw_queue_rx_dpcp.h"
 #include "dev/doca/hw_queue_rx_doca.h"
 #include "dev/net_device_table_mgr.h"
@@ -65,7 +64,8 @@ public:
     ring_simple(int if_index, ring *parent, bool use_locks);
     virtual ~ring_simple();
 
-    bool request_notification(cq_type_t cq_type) override;
+    bool request_notification_rx() override;
+    bool request_notification_tx() override;
     bool poll_and_process_element_rx(void *pv_fd_ready_array = nullptr) override;
     void poll_and_process_element_tx() override;
     void adapt_cq_moderation() override;
@@ -75,96 +75,160 @@ public:
     void mem_buf_rx_release(mem_buf_desc_t *p_mem_buf_desc) override;
     int drain_and_proccess() override;
     void clear_rx_notification() override;
-    void mem_buf_desc_return_to_owner_tx(mem_buf_desc_t *p_mem_buf_desc);
     void mem_buf_desc_return_to_owner_rx(mem_buf_desc_t *p_mem_buf_desc,
                                          void *pv_fd_ready_array = nullptr);
-    inline int send_buffer(xlio_ibv_send_wr *p_send_wqe, xlio_wr_tx_packet_attr attr,
-                           xlio_tis *tis);
     bool is_up() override;
+    mem_buf_desc_t *mem_buf_tx_get(ring_user_id_t id, pbuf_type type,
+                                   uint32_t n_num_mem_bufs = 1) override;
+    int mem_buf_tx_release(mem_buf_desc_t *p_mem_buf_desc_list, bool trylock = false) override;
+    void mem_buf_desc_return_single_to_owner_tx(mem_buf_desc_t *p_mem_buf_desc) override;
+    void mem_buf_desc_return_single_multi_ref(mem_buf_desc_t *p_mem_buf_desc,
+                                              unsigned ref) override;
+    size_t get_rx_channels_num() const override { return 1U; };
+    int get_rx_channel_fd(size_t ch_idx) const override;
+    int get_tx_channel_fd() const override;
+    uint32_t get_max_payload_sz(void) override;
+    uint16_t get_max_header_sz() override;
+    bool is_tso(void) override;
+    int modify_ratelimit(struct xlio_rate_limit_t &rate_limit) override;
+    bool tls_tx_supported() override { return m_tls.tls_tx; }
+    bool tls_rx_supported() override { return m_tls.tls_rx; }
+    ib_ctx_handler *get_ctx(ring_user_id_t) override { return m_p_ib_ctx; }
+
     void start_active_queue_tx();
     void start_active_queue_rx();
     void stop_active_queue_tx();
     void stop_active_queue_rx();
-    mem_buf_desc_t *mem_buf_tx_get(ring_user_id_t id, pbuf_type type,
-                                   uint32_t n_num_mem_bufs = 1) override;
-    int mem_buf_tx_release(mem_buf_desc_t *p_mem_buf_desc_list, bool trylock = false) override;
+    void modify_cq_moderation(uint32_t period, uint32_t count);
+    void convert_hw_time_to_system_time(uint64_t hwtime, struct timespec *systime)
+    {
+        m_p_ib_ctx->convert_hw_time_to_system_time(hwtime, systime);
+    }
+    friend class hw_queue_tx;
+    friend class hw_queue_rx;
+    friend class rfs;
+    friend class rfs_uc;
+    friend class rfs_uc_tcp_gro;
+    friend class rfs_mc;
+    friend class ring_bond;
+
+protected:
+    void create_resources();
+    virtual void init_tx_buffers(uint32_t count);
+    void inc_cq_moderation_stats() override;
+    inline uint32_t get_mtu() { return m_mtu; }
+
+private:
+    inline mem_buf_desc_t *get_tx_buffers(pbuf_type type, uint32_t n_num_mem_bufs);
+    int put_tx_buffer_helper(mem_buf_desc_t *buff);
+    inline int put_tx_buffers(mem_buf_desc_t *buff_list);
+    inline int put_tx_single_buffer(mem_buf_desc_t *buff);
+    void return_to_global_pool();
+    bool request_more_tx_buffers(pbuf_type type, uint32_t count);
+
+    void save_l2_address(const L2_address *p_l2_addr)
+    {
+        delete_l2_address();
+        m_p_l2_addr = p_l2_addr->clone();
+    };
+    void delete_l2_address()
+    {
+        if (m_p_l2_addr) {
+            delete m_p_l2_addr;
+        }
+        m_p_l2_addr = nullptr;
+    };
+
+protected:
+    ib_ctx_handler *m_p_ib_ctx;
+    hw_queue_tx *m_hqtx = nullptr;
+    hw_queue_rx *m_hqrx = nullptr;
+    struct cq_moderation_info m_cq_moderation_info;
+
+private:
+    lock_mutex m_lock_ring_tx_buf_wait;
+    uint32_t m_tx_num_bufs = 0U;
+    uint32_t m_zc_num_bufs = 0U;
+
+    gro_mgr m_gro_mgr;
+    bool m_up_tx = false;
+    bool m_up_rx = false;
+
+    L2_address *m_p_l2_addr = nullptr;
+    uint32_t m_mtu;
+
+    struct {
+        /* Maximum length of TCP payload for TSO */
+        uint32_t max_payload_sz;
+
+        /* Maximum length of header for TSO */
+        uint16_t max_header_sz;
+    } m_tso;
+    struct {
+        /* TLS TX offload is supported */
+        bool tls_tx;
+        /* TLS RX offload is supported */
+        bool tls_rx;
+        /* TLS DEK modify Crypto-Sync is supported */
+        bool tls_synchronize_dek;
+    } m_tls;
+    struct {
+        /* Indicates LRO support */
+        bool cap;
+
+        /* Indicate LRO support for segments with PSH flag */
+        bool psh_flag;
+
+        /* Indicate LRO support for segments with TCP timestamp option */
+        bool time_stamp;
+
+        /* The maximum message size mode
+         * 0x0 - TCP header + TCP payload
+         * 0x1 - L2 + L3 + TCP header + TCP payload
+         */
+        uint8_t max_msg_sz_mode;
+
+        /* The minimal size of TCP segment required for coalescing */
+        uint16_t min_mss_size;
+
+        /* Array of supported LRO timer periods in microseconds. */
+        uint8_t timer_supported_periods[4];
+
+        /* Maximum length of TCP payload for LRO
+         * It is calculated from max_msg_sz_mode and safe_mce_sys().rx_buf_size
+         */
+        uint32_t max_payload_sz;
+    } m_lro;
+
+#ifdef DEFINED_DPCP_PATH_RX
+public:
+    friend class cq_mgr_rx;
+    friend class cq_mgr_rx_regrq;
+    friend class cq_mgr_rx_strq;
+
+protected:
+    cq_mgr_rx *m_p_cq_mgr_rx = nullptr;
+    struct ibv_comp_channel *m_p_rx_comp_event_channel = nullptr;
+#endif // DEFINED_DPCP_PATH_RX
+
+#ifdef DEFINED_DPCP_PATH_TX
+public:
+    friend class cq_mgr_tx;
+    uint32_t get_max_send_sge(void) override;
+    uint32_t get_tx_user_lkey(void *addr, size_t length) override;
+    uint32_t get_max_inline_data() override;
+    bool get_hw_dummy_send_support(ring_user_id_t id, xlio_ibv_send_wr *p_send_wqe) override;
     void send_ring_buffer(ring_user_id_t id, xlio_ibv_send_wr *p_send_wqe,
                           xlio_wr_tx_packet_attr attr) override;
     int send_lwip_buffer(ring_user_id_t id, xlio_ibv_send_wr *p_send_wqe,
                          xlio_wr_tx_packet_attr attr, xlio_tis *tis) override;
-    uint32_t send_doca_single(void *ptr, uint32_t len, mem_buf_desc_t *buff) override;
-    uint32_t send_doca_lso(struct iovec &h, struct pbuf *p, uint16_t mss,
-                           bool is_zerocopy) override;
-    void mem_buf_desc_return_single_to_owner_tx(mem_buf_desc_t *p_mem_buf_desc) override;
-    void mem_buf_desc_return_single_multi_ref(mem_buf_desc_t *p_mem_buf_desc,
-                                              unsigned ref) override;
-    bool get_hw_dummy_send_support(ring_user_id_t id, xlio_ibv_send_wr *p_send_wqe) override;
-    inline void convert_hw_time_to_system_time(uint64_t hwtime, struct timespec *systime)
-    {
-        m_p_ib_ctx->convert_hw_time_to_system_time(hwtime, systime);
-    }
-    int modify_ratelimit(struct xlio_rate_limit_t &rate_limit) override;
-    size_t get_rx_channels_num() const override { return 1U; };
-    int get_rx_channel_fd(size_t ch_idx) const override;
-    int get_tx_channel_fd() const override;
-    uint32_t get_tx_user_lkey(void *addr, size_t length) override;
-    uint32_t get_max_inline_data() override;
-    ib_ctx_handler *get_ctx(ring_user_id_t id) override
-    {
-        NOT_IN_USE(id);
-        return m_p_ib_ctx;
-    }
-    uint32_t get_max_send_sge(void) override;
-    uint32_t get_max_payload_sz(void) override;
-    uint16_t get_max_header_sz(void) override;
     uint32_t get_tx_lkey(ring_user_id_t id) override
     {
         NOT_IN_USE(id);
         return m_tx_lkey;
     }
-    bool is_tso() override;
-
-    void modify_cq_moderation(uint32_t period, uint32_t count);
-
-#ifdef DEFINED_UTLS
-    bool tls_tx_supported(void) override { return m_tls.tls_tx; }
-    bool tls_rx_supported(void) override { return m_tls.tls_rx; }
-    bool tls_sync_dek_supported() { return m_tls.tls_synchronize_dek; }
-    xlio_tis *tls_context_setup_tx(const xlio_tls_info *info) override
-    {
-        std::lock_guard<decltype(m_lock_ring_tx)> lock(m_lock_ring_tx);
-
-        xlio_tis *tis = m_hqtx->tls_context_setup_tx(info);
-        if (likely(tis != NULL)) {
-            ++m_p_ring_stat->n_tx_tls_contexts;
-        }
-
-        /* Do polling to speedup handling of the completion. */
-        m_p_cq_mgr_tx->poll_and_process_element_tx();
-
-        return tis;
-    }
-    void tls_context_resync_tx(const xlio_tls_info *info, xlio_tis *tis, bool skip_static) override
-    {
-        std::lock_guard<decltype(m_lock_ring_tx)> lock(m_lock_ring_tx);
-        m_hqtx->tls_context_resync_tx(info, tis, skip_static);
-        m_p_cq_mgr_tx->poll_and_process_element_tx();
-    }
-    void tls_release_tis(xlio_tis *tis) override
-    {
-        std::lock_guard<decltype(m_lock_ring_tx)> lock(m_lock_ring_tx);
-        m_hqtx->tls_release_tis(tis);
-    }
-    void tls_tx_post_dump_wqe(xlio_tis *tis, void *addr, uint32_t len, uint32_t lkey,
-                              bool first) override
-    {
-        std::lock_guard<decltype(m_lock_ring_tx)> lock(m_lock_ring_tx);
-        if (lkey == LKEY_TX_DEFAULT) {
-            lkey = m_tx_lkey;
-        }
-        m_hqtx->tls_tx_post_dump_wqe(tis, addr, len, lkey, first);
-    }
-#endif /* DEFINED_UTLS */
+    inline int send_buffer(xlio_ibv_send_wr *p_send_wqe, xlio_wr_tx_packet_attr attr,
+                           xlio_tis *tis);
 
     std::unique_ptr<xlio_tis> create_tis(uint32_t flags) const override
     {
@@ -204,118 +268,67 @@ public:
         m_hqtx->credits_return(credits);
     }
 
-    friend class cq_mgr_tx;
-    friend class hw_queue_tx;
-    friend class hw_queue_rx;
-    friend class rfs;
-    friend class rfs_uc;
-    friend class rfs_uc_tcp_gro;
-    friend class rfs_mc;
-    friend class ring_bond;
-
 protected:
-    void create_resources();
-    virtual void init_tx_buffers(uint32_t count);
-    void inc_cq_moderation_stats() override;
-    inline void set_tx_num_wr(uint32_t num_wr) { m_tx_num_wr = num_wr; }
-    inline uint32_t get_tx_num_wr() { return m_tx_num_wr; }
-    inline uint32_t get_mtu() { return m_mtu; }
-
-private:
-    inline void send_status_handler(int ret, xlio_ibv_send_wr *p_send_wqe);
-    int put_tx_buffer_helper(mem_buf_desc_t *buff);
-    inline int put_tx_buffers(mem_buf_desc_t *buff_list);
-    inline int put_tx_single_buffer(mem_buf_desc_t *buff);
-    void return_to_global_pool();
-    bool is_available_qp_wr(bool b_block, unsigned credits);
-    void save_l2_address(const L2_address *p_l2_addr)
-    {
-        delete_l2_address();
-        m_p_l2_addr = p_l2_addr->clone();
-    };
-    void delete_l2_address()
-    {
-        if (m_p_l2_addr) {
-            delete m_p_l2_addr;
-        }
-        m_p_l2_addr = nullptr;
-    };
-
-protected:
-    ib_ctx_handler *m_p_ib_ctx;
-    hw_queue_tx *m_hqtx = nullptr;
-    hw_queue_rx *m_hqrx = nullptr;
-    struct cq_moderation_info m_cq_moderation_info;
-    cq_mgr_tx *m_p_cq_mgr_tx = nullptr;
     std::unordered_map<void *, uint32_t> m_user_lkey_map;
-
-private:
-    lock_mutex m_lock_ring_tx_buf_wait;
-    uint32_t m_tx_num_bufs = 0U;
-    uint32_t m_zc_num_bufs = 0U;
+    cq_mgr_tx *m_p_cq_mgr_tx = nullptr;
+    struct ibv_comp_channel *m_p_tx_comp_event_channel = nullptr;
     uint32_t m_tx_num_wr = 0U;
     uint32_t m_tx_lkey = 0U; // this is the registered memory lkey for a given specific device for
                              // the buffer pool use
-    doca_mmap *m_p_doca_mmap;
-    gro_mgr m_gro_mgr;
-    bool m_up_tx = false;
-    bool m_up_rx = false;
-    struct ibv_comp_channel *m_p_rx_comp_event_channel = nullptr;
-    struct ibv_comp_channel *m_p_tx_comp_event_channel = nullptr;
-    L2_address *m_p_l2_addr = nullptr;
-    uint32_t m_mtu;
-
-    struct {
-        /* Maximum length of TCP payload for TSO */
-        uint32_t max_payload_sz;
-
-        /* Maximum length of header for TSO */
-        uint16_t max_header_sz;
-    } m_tso;
-#ifdef DEFINED_UTLS
-    struct {
-        /* TLS TX offload is supported */
-        bool tls_tx;
-        /* TLS RX offload is supported */
-        bool tls_rx;
-        /* TLS DEK modify Crypto-Sync is supported */
-        bool tls_synchronize_dek;
-    } m_tls;
-#endif /* DEFINED_UTLS */
-    struct {
-        /* Indicates LRO support */
-        bool cap;
-
-        /* Indicate LRO support for segments with PSH flag */
-        bool psh_flag;
-
-        /* Indicate LRO support for segments with TCP timestamp option */
-        bool time_stamp;
-
-        /* The maximum message size mode
-         * 0x0 - TCP header + TCP payload
-         * 0x1 - L2 + L3 + TCP header + TCP payload
-         */
-        uint8_t max_msg_sz_mode;
-
-        /* The minimal size of TCP segment required for coalescing */
-        uint16_t min_mss_size;
-
-        /* Array of supported LRO timer periods in microseconds. */
-        uint8_t timer_supported_periods[4];
-
-        /* Maximum length of TCP payload for LRO
-         * It is calculated from max_msg_sz_mode and safe_mce_sys().rx_buf_size
-         */
-        uint32_t max_payload_sz;
-    } m_lro;
-
-#ifdef DEFINED_DPCP_PATH_RX
+private:
+    inline void send_status_handler(int ret, xlio_ibv_send_wr *p_send_wqe);
+    bool is_available_qp_wr(bool b_block, unsigned credits);
+#else // DEFINED_DPCP_PATH_TX
 public:
-    friend class cq_mgr_rx;
-    friend class cq_mgr_rx_regrq;
-    friend class cq_mgr_rx_strq;
-#ifdef DEFINED_UTLS
+    uint32_t send_doca_single(void *ptr, uint32_t len, mem_buf_desc_t *buff) override;
+    uint32_t send_doca_lso(struct iovec &h, struct pbuf *p, uint16_t mss,
+                           bool is_zerocopy) override;
+
+private:
+    doca_mmap *m_p_doca_mmap;
+#endif // DEFINED_DPCP_PATH_TX
+
+#if defined(DEFINED_DPCP_PATH_TX) && defined(DEFINED_UTLS)
+public:
+    bool tls_sync_dek_supported() { return m_tls.tls_synchronize_dek; }
+    xlio_tis *tls_context_setup_tx(const xlio_tls_info *info) override
+    {
+        std::lock_guard<decltype(m_lock_ring_tx)> lock(m_lock_ring_tx);
+
+        xlio_tis *tis = m_hqtx->tls_context_setup_tx(info);
+        if (likely(tis != NULL)) {
+            ++m_p_ring_stat->n_tx_tls_contexts;
+        }
+
+        /* Do polling to speedup handling of the completion. */
+        m_p_cq_mgr_tx->poll_and_process_element_tx();
+
+        return tis;
+    }
+    void tls_context_resync_tx(const xlio_tls_info *info, xlio_tis *tis, bool skip_static) override
+    {
+        std::lock_guard<decltype(m_lock_ring_tx)> lock(m_lock_ring_tx);
+        m_hqtx->tls_context_resync_tx(info, tis, skip_static);
+        m_p_cq_mgr_tx->poll_and_process_element_tx();
+    }
+    void tls_release_tis(xlio_tis *tis) override
+    {
+        std::lock_guard<decltype(m_lock_ring_tx)> lock(m_lock_ring_tx);
+        m_hqtx->tls_release_tis(tis);
+    }
+    void tls_tx_post_dump_wqe(xlio_tis *tis, void *addr, uint32_t len, uint32_t lkey,
+                              bool first) override
+    {
+        std::lock_guard<decltype(m_lock_ring_tx)> lock(m_lock_ring_tx);
+        if (lkey == LKEY_TX_DEFAULT) {
+            lkey = m_tx_lkey;
+        }
+        m_hqtx->tls_tx_post_dump_wqe(tis, addr, len, lkey, first);
+    }
+#endif // DEFINED_DPCP_PATH_TX && DEFINED_UTLS
+
+#if defined(DEFINED_DPCP_PATH_RX_AND_TX) && defined(DEFINED_UTLS)
+public:
     xlio_tir *tls_create_tir(bool cached) override
     {
         /*
@@ -363,10 +376,7 @@ public:
         std::lock_guard<decltype(m_lock_ring_tx)> lock(m_lock_ring_tx);
         m_hqrx->tls_release_tir(tir);
     }
-#endif // DEFINED_UTLS
-protected:
-    cq_mgr_rx *m_p_cq_mgr_rx = nullptr;
-#endif // DEFINED_DPCP_PATH_RX
+#endif // DEFINED_DPCP_PATH_RX && DEFINED_DPCP_PATH_TX && DEFINED_UTLS
 };
 
 class ring_eth : public ring_simple {
