+++ src/core/proto/neighbour.cpp
@@ -546,7 +546,9 @@ bool neigh_entry::post_send_udp_ipv4(neigh_send_data *n_send_data)
         m_sge.addr =
             (uintptr_t)(p_mem_buf_desc->p_buffer + (uint8_t)h->m_transport_header_tx_offset);
         m_sge.length = sz_user_data_to_copy + hdr_len;
+#ifdef DEFINED_DPCP_PATH_TX
         m_sge.lkey = m_p_ring->get_tx_lkey(m_id);
+#endif // DEFINED_DPCP_PATH_TX
         m_send_wqe.wr_id = (uintptr_t)p_mem_buf_desc;
 
         neigh_logdbg("packet_sz=%d, payload_sz=%zd, ip_offset=%d id=%d",
@@ -559,8 +561,9 @@ bool neigh_entry::post_send_udp_ipv4(neigh_send_data *n_send_data)
 
         // We don't check the return value of post send when we reach the HW we consider that we
         // completed our job
+#ifdef DEFINED_DPCP_PATH_TX
         m_p_ring->send_ring_buffer(m_id, &m_send_wqe, XLIO_TX_PACKET_L3_CSUM);
-
+#endif // DEFINED_DPCP_PATH_TX
         p_mem_buf_desc = tmp;
 
         // Update ip frag offset position
@@ -589,10 +592,15 @@ bool neigh_entry::post_send_udp_ipv6_fragmented(neigh_send_data *n_send_data, si
         return false;
     }
 
+#ifdef DEFINED_DPCP_PATH_TX
     return dst_entry_udp::fast_send_fragmented_ipv6(
         p_mem_buf_desc, &n_send_data->m_iov, 1, XLIO_TX_PACKET_L3_CSUM, sz_udp_payload, n_num_frags,
         &m_send_wqe, m_id, &m_sge, n_send_data->m_header, max_ip_payload_size, m_p_ring,
         n_send_data->m_packet_id);
+#else
+    NOT_IN_USE(n_send_data);
+    return 0;
+#endif // DEFINED_DPCP_PATH_TX
 }
 
 bool neigh_entry::post_send_udp_ipv6_not_fragmented(neigh_send_data *n_send_data)
@@ -643,13 +651,11 @@ bool neigh_entry::post_send_udp_ipv6_not_fragmented(neigh_send_data *n_send_data
     }
     BULLSEYE_EXCLUDE_BLOCK_END
 
-    wqe_send_handler wqe_sh;
-    xlio_wr_tx_packet_attr attr =
-        (xlio_wr_tx_packet_attr)(XLIO_TX_PACKET_L4_CSUM | XLIO_TX_PACKET_L3_CSUM);
-
     m_sge.addr = (uintptr_t)(p_mem_buf_desc->p_buffer + (uint8_t)h->m_transport_header_tx_offset);
     m_sge.length = sz_data_payload + hdr_len;
+#ifdef DEFINED_DPCP_PATH_TX
     m_sge.lkey = m_p_ring->get_tx_lkey(m_id);
+#endif // DEFINED_DPCP_PATH_TX
     m_send_wqe.wr_id = reinterpret_cast<uintptr_t>(p_mem_buf_desc);
 
     neigh_logdbg("packet_sz=%d, payload_sz=%zd, id=%d", m_sge.length - h->m_transport_header_len,
@@ -657,8 +663,11 @@ bool neigh_entry::post_send_udp_ipv6_not_fragmented(neigh_send_data *n_send_data
 
     // We don't check the return value of post send when we reach the HW we consider that we
     // completed our job
+#ifdef DEFINED_DPCP_PATH_TX
+    xlio_wr_tx_packet_attr attr =
+        (xlio_wr_tx_packet_attr)(XLIO_TX_PACKET_L4_CSUM | XLIO_TX_PACKET_L3_CSUM);
     m_p_ring->send_ring_buffer(m_id, &m_send_wqe, attr);
-
+#endif // DEFINED_DPCP_PATH_TX
     return true;
 }
 
@@ -707,8 +716,9 @@ bool neigh_entry::post_send_tcp(neigh_send_data *p_data)
     size_t hdr_alignment_diff = h->m_aligned_l2_l3_len - h->m_total_hdr_len;
     m_sge.addr = (uintptr_t)((uint8_t *)p_pkt + hdr_alignment_diff);
     m_sge.length = total_packet_len;
+#ifdef DEFINED_DPCP_PATH_TX
     m_sge.lkey = m_p_ring->get_tx_lkey(m_id);
-
+#endif // DEFINED_DPCP_PATH_TX
     /* for DEBUG */
     if ((uint8_t *)m_sge.addr < p_mem_buf_desc->p_buffer) {
         neigh_logerr("p_buffer - addr=%d, m_total_hdr_len=%u, p_buffer=%p, type=%d, len=%d, "
@@ -720,17 +730,16 @@ bool neigh_entry::post_send_tcp(neigh_send_data *p_data)
     }
 
     m_send_wqe.wr_id = (uintptr_t)p_mem_buf_desc;
-    xlio_wr_tx_packet_attr attr =
-        (xlio_wr_tx_packet_attr)(XLIO_TX_PACKET_L3_CSUM | XLIO_TX_PACKET_L4_CSUM);
     p_mem_buf_desc->tx.p_ip_h = p_ip_hdr;
     p_mem_buf_desc->tx.p_tcp_h = reinterpret_cast<tcphdr *>(p_tcp_hdr);
 
-    if (!safe_mce_sys().doca_tx) {
-        m_p_ring->send_ring_buffer(m_id, &m_send_wqe, attr);
-    } else {
-        m_p_ring->send_doca_single(reinterpret_cast<void *>(m_sge.addr), m_sge.length,
-                                   p_mem_buf_desc);
-    }
+#ifdef DEFINED_DPCP_PATH_TX
+    xlio_wr_tx_packet_attr attr =
+        (xlio_wr_tx_packet_attr)(XLIO_TX_PACKET_L3_CSUM | XLIO_TX_PACKET_L4_CSUM);
+    m_p_ring->send_ring_buffer(m_id, &m_send_wqe, attr);
+#else // DEFINED_DPCP_PATH_TX
+    m_p_ring->send_doca_single(reinterpret_cast<void *>(m_sge.addr), m_sge.length, p_mem_buf_desc);
+#endif // DEFINED_DPCP_PATH_TX
 
 #ifndef __COVERITY__
     struct tcphdr *p_tcp_h = reinterpret_cast<tcphdr *>(p_tcp_hdr);
@@ -1666,12 +1675,11 @@ bool neigh_eth::send_arp_request(bool is_broadcast)
     p_mem_buf_desc->p_next_desc = nullptr;
     m_send_wqe.wr_id = (uintptr_t)p_mem_buf_desc;
 
-    if (!safe_mce_sys().doca_tx) {
-        m_p_ring->send_ring_buffer(m_id, &m_send_wqe, (xlio_wr_tx_packet_attr)0);
-    } else {
-        m_p_ring->send_doca_single(reinterpret_cast<void *>(m_sge.addr), m_sge.length,
-                                   p_mem_buf_desc);
-    }
+#ifdef DEFINED_DPCP_PATH_TX
+    m_p_ring->send_ring_buffer(m_id, &m_send_wqe, (xlio_wr_tx_packet_attr)0);
+#else // DEFINED_DPCP_PATH_TX
+    m_p_ring->send_doca_single(reinterpret_cast<void *>(m_sge.addr), m_sge.length, p_mem_buf_desc);
+#endif // DEFINED_DPCP_PATH_TX
 
     neigh_logdbg("ARP Sent");
     return true;
@@ -1805,12 +1813,11 @@ bool neigh_eth::send_neighbor_solicitation()
     neigh_logdbg("NS request: base=%p addr=%p length=%" PRIu32, p_mem_buf_desc->p_buffer,
                  (void *)m_sge.addr, m_sge.length);
 
-    if (!safe_mce_sys().doca_tx) {
-        m_p_ring->send_ring_buffer(m_id, &m_send_wqe, (xlio_wr_tx_packet_attr)0);
-    } else {
-        m_p_ring->send_doca_single(reinterpret_cast<void *>(m_sge.addr), m_sge.length,
-                                   p_mem_buf_desc);
-    }
+#ifdef DEFINED_DPCP_PATH_TX
+    m_p_ring->send_ring_buffer(m_id, &m_send_wqe, (xlio_wr_tx_packet_attr)0);
+#else // DEFINED_DPCP_PATH_TX
+    m_p_ring->send_doca_single(reinterpret_cast<void *>(m_sge.addr), m_sge.length, p_mem_buf_desc);
+#endif // DEFINED_DPCP_PATH_TX
 
     neigh_logdbg("Neighbor solicitation has been sent");
     return true;
