+++ src/core/sock/sockinfo_udp.cpp
@@ -1993,7 +1993,7 @@ void sockinfo_udp::rx_request_notification()
     for (rx_ring_map_t::iterator rx_ring_iter = m_rx_ring_map.begin();
          rx_ring_iter != m_rx_ring_map.end(); rx_ring_iter++) {
         ring *p_ring = rx_ring_iter->first;
-        if (!p_ring->request_notification(CQT_RX)) {
+        if (!p_ring->request_notification_rx()) {
             si_udp_logerr("Failure from ring[%p]->request_notification()", p_ring);
         }
     }
@@ -2125,6 +2125,7 @@ ssize_t sockinfo_udp::tx(xlio_tx_call_attr_t &tx_arg)
     }
 
     {
+#ifdef DEFINED_DPCP_PATH_TX
         xlio_send_attr attr = {(xlio_wr_tx_packet_attr)0, 0, 0, nullptr};
         bool b_blocking = m_b_blocking;
         if (unlikely(__flags & MSG_DONTWAIT)) {
@@ -2136,12 +2137,14 @@ ssize_t sockinfo_udp::tx(xlio_tx_call_attr_t &tx_arg)
                                               (is_dummy * XLIO_TX_PACKET_DUMMY));
         if (likely(p_dst_entry->is_valid())) {
             // All set for fast path packet sending - this is our best performance flow
+
             ret = p_dst_entry->fast_send(p_iov, sz_iov, attr);
         } else {
             // updates the dst_entry internal information and packet headers
             ret = p_dst_entry->slow_send(p_iov, sz_iov, attr, m_so_ratelimit, __flags, this,
                                          tx_arg.opcode);
         }
+#endif // DEFINED_DPCP_PATH_TX
 
         // Condition for cache optimization
         if (unlikely(safe_mce_sys().ring_migration_ratio_tx > 0)) {
